% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{add_cluster_label}
\alias{add_cluster_label}
\title{Add cluster label near beginning of plot (only for 1d data).}
\usage{
add_cluster_label(res, cex = 1.5)
}
\arguments{
\item{cex}{if NULL, don't make cluster label.}
}
\description{
Add cluster label near beginning of plot (only for 1d data).
}
