% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d-extra.R
\name{add_date_ticks}
\alias{add_date_ticks}
\title{Make ticks from rownames of res$X. TODO: make it handle dates. (only for 1d
data).}
\usage{
add_date_ticks(res)
}
\arguments{
\item{res}{Object of class |flowmix|.}
}
\description{
Make ticks from rownames of res$X. TODO: make it handle dates. (only for 1d
data).
}
