% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{add_date_ticks_from_dates}
\alias{add_date_ticks_from_dates}
\title{Add date ticks from string of dates.}
\usage{
add_date_ticks_from_dates(dates, empty_tick_marks = FALSE, ...)
}
\arguments{
\item{dates}{Vector of strings of the form "2017-05-29T00:00:00", or
otherwise recognizeable using \code{lubridate::as_datetime()}.}

\item{empty_tick_marks}{Defaults to FALSE. If TRUE, use EMPTY tick marks.}

\item{...}{Rest of arguments to both axes via \code{axis()}.}
}
\value{
No return.
}
\description{
Add date ticks from string of dates.
}
