% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette-helpers.R
\name{add_model_2d}
\alias{add_model_2d}
\title{Add to 2d plot the estimated cluster mean and 95% confidence region, for
each cluster.}
\usage{
add_model_2d(p, mn, sigma, prob)
}
\arguments{
\item{p}{ggplot of 2d scatterplot of data}

\item{mn}{mean}

\item{sigma}{covariance}

\item{prob}{relative abundance}
}
\description{
Add to 2d plot the estimated cluster mean and 95% confidence region, for
each cluster.
}
\keyword{internal}
