% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data-flowcut.R, R/generate-data.R
\name{add_transition}
\alias{add_transition}
\title{Add a column of 0's and 1's with one changepoint, demarcated by transition
region crossings.}
\usage{
add_transition(X, lat)

add_transition(X, lat)
}
\arguments{
\item{X}{Covariates.}

\item{lat}{Latitude.}
}
\description{
Add a column of 0's and 1's with one changepoint, demarcated by transition
region crossings.

Add a column of 0's and 1's with one changepoint, demarcated by transition
region crossings.
}
