% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R
\name{admm_oneclust}
\alias{admm_oneclust}
\title{ADMM of one cluster.}
\usage{
admm_oneclust(
  iclust,
  niter,
  Xtilde,
  yvec,
  p,
  TT,
  N,
  dimdat,
  maxdev,
  Xa,
  rho,
  rhoinit,
  Xinv,
  schurA,
  schurB,
  term3,
  sigmainv,
  Xaug,
  ybar,
  Q,
  lambda,
  resp,
  resp.sum,
  ylist,
  X,
  tX,
  err_rel,
  err_abs,
  zerothresh,
  beta,
  Z,
  W,
  U,
  first_iter,
  outer_iter,
  local_adapt,
  sigma,
  sigma_eig_by_clust,
  space = 20
)
}
\arguments{
\item{local_adapt}{TRUE if locally adaptive ADMM is to be used.}
}
\value{
List containing |beta|, |yhat|, |resid_mat|, |fits|.
}
\description{
ADMM of one cluster.
}
