% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alter_beta}
\alias{alter_beta}
\title{Altering the beta coefficient (for all clusters) by adding zero coefficients
to certain rows.}
\usage{
alter_beta(beta, flatX, orig_names)
}
\arguments{
\item{beta}{beta coefficients (list of p x dimdat matrices).}

\item{flatX}{Indices, out of 1 through p, of which variables are flat.}

\item{orig_names}{Original column names of X.}
}
\value{
Same format as beta, but all-zero rows added to beta matrix in each
  cluster.
}
\description{
Altering the beta coefficient (for all clusters) by adding zero coefficients
to certain rows.
}
