% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alter_W}
\alias{alter_W}
\title{Altering the hidden layer weights by adding zero coefficients
to certain rows.}
\usage{
alter_W(W, flatX, orig_names)
}
\arguments{
\item{W}{Hidden layer weights ((p + 1) x n_hidden_nodes matrix). Assumes first 
column corresponds to the intercept/bias term.}

\item{flatX}{Indices, out of 1 through p, of which variables are flat.}

\item{orig_names}{Original column names of X.}
}
\value{
Same format as W, but with zeros in the columns where variable
  that was flat.
}
\description{
Altering the hidden layer weights by adding zero coefficients
to certain rows.
}
