% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-data.R
\name{bin_many_cytograms_2d}
\alias{bin_many_cytograms_2d}
\title{Main function for binning many cytograms (\code{ylist}). Only works for
\code{ylist} whose dimension is 2.}
\usage{
bin_many_cytograms_2d(
  ylist,
  manual.grid,
  verbose = FALSE,
  mc.cores = 1,
  qclist = NULL
)
}
\arguments{
\item{ylist}{\code{TT} lengthed list of (\code{nt} by \code{dimdat})
matrices.}

\item{manual.grid}{grid, produced using \code{make_grid()}.}

\item{verbose}{TRUE for loudness.}

\item{mc.cores}{Use for multiple-core calculations.}

\item{qclist}{Biomass (Qc) of each particle. Defaults to NULL.}
}
\value{
List containing *trimmed* ybin (a x 3) and counts (a).
}
\description{
Main function for binning many cytograms (\code{ylist}). Only works for
\code{ylist} whose dimension is 2.
}
