% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-data.R
\name{bin_one_cytogram}
\alias{bin_one_cytogram}
\title{Main function for binning a cytogram (y). Only works for y whose dimension is 3.}
\usage{
bin_one_cytogram(y, manual.grid, qc = NULL)
}
\arguments{
\item{y}{nt by dimdat matrix.}

\item{manual.grid}{grid, produced using \code{make_grid()}.}

\item{qc}{QC value, but in the original scale, NOT in the log scale.}
}
\value{
List containing *trimmed* ybin (a x 3) and counts (a).
}
\description{
Main function for binning a cytogram (y). Only works for y whose dimension is 3.
}
