% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{bin_plot_1d}
\alias{bin_plot_1d}
\title{Make a 1d binned data plot, using ggplot.}
\usage{
bin_plot_1d(ylist, qclist, col = NULL)
}
\arguments{
\item{ylist}{Must be numeric vectors containing bin centers.}

\item{qclist}{Corresponding QC in those bins.}

\item{col}{Vector of colors.}
}
\value{
ggplot object.
}
\description{
Make a 1d binned data plot, using ggplot.
}
