% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2d.R
\name{bin_plot_2d}
\alias{bin_plot_2d}
\title{2d Plotting functionality using ggplot2 (only 2d data).}
\usage{
bin_plot_2d(
  datobj_2d,
  mn = NULL,
  sigma = NULL,
  prob = NULL,
  labels = NULL,
  fac = 20,
  colours = NULL,
  mn_colours = NULL
)
}
\arguments{
\item{datobj_2d}{A data matrix with X on grid; three columns are assumed;
the first two are the coordinates of the 2d data grid (e.g. "diam" and
"chl"); the thrid column is named "counts".}
}
\value{
A ggplot object.
}
\description{
2d Plotting functionality using ggplot2 (only 2d data).
}
