% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-proximal.R
\name{calc_objective}
\alias{calc_objective}
\title{Objective function calculation
fit is T x d, Y_c_wt is d x T
We need W * (Y_unweighted - fit)^2
Which is equivalent to (Y_weighted - W * fit)^2 / W}
\usage{
calc_objective(X_c, Y_c_wt, W, beta, N_w, res, sigmainv)
}
\description{
Objective function calculation
fit is T x d, Y_c_wt is d x T
We need W * (Y_unweighted - fit)^2
Which is equivalent to (Y_weighted - W * fit)^2 / W
}
