% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{check_compatible}
\alias{check_compatible}
\title{Checks whether the dataset and \code{flowmix}-class object are
compatible. (Under construction! Can you think of anything else?). Used in
unit tests.}
\usage{
check_compatible(ylist, res)
}
\arguments{
\item{ylist}{Data.}

\item{res}{Model; \code{flowmix} class object.}
}
\description{
Checks whether the dataset and \code{flowmix}-class object are
compatible. (Under construction! Can you think of anything else?). Used in
unit tests.
}
