% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{check_if_date}
\alias{check_if_date}
\title{Check if time is TRUE.}
\usage{
check_if_date(date)
}
\arguments{
\item{date}{String vector containing date strings
e.g. "2017-06-13T11:00:00".}
}
\value{
TRUE only all are dates..
}
\description{
Check if time is TRUE.
}
