% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{check_if_same_size}
\alias{check_if_same_size}
\title{Checks whether two objects of \code{flowmix} class are exactly the same
size.}
\usage{
check_if_same_size(ylist1, ylist2)
}
\arguments{
\item{ylist1}{One \code{flowmix} class object.}

\item{ylist2}{Another \code{flowmix} class object.}
}
\description{
Checks whether two objects of \code{flowmix} class are exactly the same
size.
}
