% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_zero_stabilize}
\alias{check_zero_stabilize}
\title{Check if zero pattern in coefficients, across EM iterations, has stabilized.}
\usage{
check_zero_stabilize(zero.betas, zero.alphas, iter)
}
\arguments{
\item{zero.betas}{List of patterns in beta, over EM interations.}

\item{zero.alphas}{List of zero patterns in alpha, over EM iterations.}

\item{iter}{Iteration number.}
}
\value{
TRUE if both beta and alpha coefficients' zero patterns have
  stabilized.
}
\description{
Check if zero pattern in coefficients, across EM iterations, has stabilized.
}
