% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{choldecomp_sigma}
\alias{choldecomp_sigma}
\title{Gets cholesky decomposition.}
\usage{
choldecomp_sigma(sigma)
}
\arguments{
\item{sigma}{A single (dimdat x dimdat) matrix.}
}
\value{
List containing eigenvalues (vector), eigenvectors (matrix of
  eigenvectors as columns)
}
\description{
Gets cholesky decomposition.
}
