% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{choldecomp_sigma_array}
\alias{choldecomp_sigma_array}
\title{From a (TT x numclust x dimdat x dimdat) array whose [tt,ii,,]'th entry is
the (dimdat x dimdat) covariance matrix, do cholesky decompositions.}
\usage{
choldecomp_sigma_array(sigma_array)
}
\arguments{
\item{sigma_array}{(TT x numclust x dimdat x dimdat) array}
}
\value{
TT length list of (numclust length list of eigendecompositions).
}
\description{
From a (TT x numclust x dimdat x dimdat) array whose [tt,ii,,]'th entry is
the (dimdat x dimdat) covariance matrix, do cholesky decompositions.
}
