% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{collapse_3d_to_1d}
\alias{collapse_3d_to_1d}
\title{Collapse 3d data to 1d data.}
\usage{
collapse_3d_to_1d(ylist, countslist, idim)
}
\arguments{
\item{ylist}{ylist}

\item{countslist}{countslist}

\item{idim}{Which dimension to collapse to (out of the
\code{ncol(ylist[[1]])} dimensions.) Alternatively, this can be a column
name of the tables in \code{ylist}}
}
\value{
Collapsed list of ylist and countslist
}
\description{
Collapse 3d data to 1d data.
}
