% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R
\name{converge}
\alias{converge}
\title{(Helper) Check convergence for ADMM.}
\usage{
converge(
  beta1,
  rho,
  w,
  Z,
  w_prev,
  Z_prev,
  Uw,
  Uz,
  tX,
  Xbeta1,
  err_rel = 1e-04,
  err_abs = 0
)
}
\arguments{
\item{err_rel}{= 1E-3}

\item{inheritParams}{Mstep_beta_admm}
}
\description{
(Helper) Check convergence for ADMM.
}
