% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.flowcut}
\alias{cv.flowcut}
\title{Cross-validation wrapper for flowmix(). Saves results to separate files in
\code{destin}.}
\usage{
cv.flowcut(
  ylist,
  countslist,
  X,
  cens_ind_left,
  cens_ind_right,
  cens_vec_1,
  cens_vec_2,
  destin = ".",
  mean_lambdas,
  prob_lambdas,
  iimat = NULL,
  maxdev,
  numclust,
  nfold,
  nrep,
  verbose = FALSE,
  refit = FALSE,
  save_meta = FALSE,
  mc.cores = 1,
  blocksize,
  folds = NULL,
  seedtab = NULL,
  flatX_thresh = 1e-05,
  ...
)
}
\arguments{
\item{ylist}{List of responses}

\item{countslist}{Multiplicity for particles in \code{ylist}.}

\item{X}{Covariates.}

\item{cens_ind_left}{the left censoring indicators for the data.}

\item{cens_ind_right}{the right censoring indicators for the data.}

\item{cens_vec_1}{the lower censoring values.}

\item{cens_vec_2}{the upper censoring values.}

\item{destin}{Where to save the output.}

\item{mean_lambdas}{Regularization parameters for betas.}

\item{prob_lambdas}{Regularization parameters for alphas.}

\item{iimat}{Indices for the cross validation jobs -- made using
\code{make_iimat()}.}

\item{maxdev}{Maximum deviation of cluster means across time.}

\item{numclust}{Number of clusters}

\item{nfold}{Number of cross-validation folds. Defaults to 5.}

\item{nrep}{Number of repetitions.}

\item{verbose}{TRUE for loudness.}

\item{refit}{If TRUE, estimate the model on the full data, for each pair of
regularization parameters.}

\item{save_meta}{If TRUE, save meta data.}

\item{mc.cores}{Use this many CPU cores.}

\item{blocksize}{Contiguous time blocks from which to form CV time folds.}

\item{folds}{Manually provide CV folds (list of time points of data to use
as CV folds). Defaults to NULL.}

\item{seedtab}{A table containing seeds (7 columns)}

\item{flatX_thresh}{Threshold for detecting if any covariates are flat (low
variance). These flat coefficients will have be set to zero and excluded
from estimation altogether.}

\item{...}{Additional arguments to flowcut().}
}
\value{
No return.
}
\description{
Cross-validation wrapper for flowmix(). Saves results to separate files in
\code{destin}.
}
