% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.flowmix}
\alias{cv.flowmix}
\title{Cross-validation wrapper for flowmix(). Saves results to separate files in
\code{destin}.}
\usage{
cv.flowmix(
  ylist,
  countslist,
  X,
  destin = ".",
  mean_lambdas,
  prob_lambdas,
  iimat = NULL,
  maxdev,
  numclust,
  nfold,
  nrep,
  verbose = FALSE,
  refit = FALSE,
  save_meta = FALSE,
  mc.cores = 1,
  blocksize,
  folds = NULL,
  seedtab = NULL,
  flatX_thresh = 1e-05,
  ...
)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{countslist}{Multiplicity for particles in \code{ylist}.}

\item{X}{Matrix of size (T x p+1)}

\item{destin}{Where to save the output.}

\item{mean_lambdas}{Regularization parameters for betas.}

\item{prob_lambdas}{Regularization parameters for alphas.}

\item{iimat}{Indices for the cross validation jobs -- made using
\code{make_iimat()}.}

\item{maxdev}{Maximum deviation of cluster means across time.}

\item{numclust}{Number of clusters}

\item{nfold}{Number of cross-validation folds. Defaults to 5.}

\item{nrep}{Number of repetitions.}

\item{verbose}{TRUE for loudness.}

\item{refit}{If TRUE, estimate the model on the full data, for each pair of
regularization parameters.}

\item{save_meta}{If TRUE, save meta data.}

\item{mc.cores}{Use this many CPU cores. If greater than 1 and R is 
linked to a multi-threaded BLAS, user should set BLAS to run with a single thread 
to avoid oversubscribing cores. This can be done by running 
\code{RhpcBLASctl::blas_set_num_threads(1); RhpcBLASctl::omp_set_num_threads(1)}.}

\item{blocksize}{Contiguous time blocks from which to form CV time folds.}

\item{folds}{Manually provide CV folds (list of time points of data to use
as CV folds). Defaults to NULL.}

\item{seedtab}{A table containing seeds (7 columns)}

\item{flatX_thresh}{Threshold for detecting if any covariates are flat (low
variance). These flat coefficients will have be set to zero and excluded
from estimation altogether.}

\item{...}{Additional arguments to flowmix().}
}
\value{
No return.
}
\description{
Cross-validation wrapper for flowmix(). Saves results to separate files in
\code{destin}.
}
\details{
If this function is run in parallel (\code{mc.cores > 1}) and R is 
linked to a multi-threaded BLAS, user should set BLAS to run with a single thread 
to avoid oversubscribing cores. This can be done by running 
\code{RhpcBLASctl::blas_set_num_threads(1); RhpcBLASctl::omp_set_num_threads(1)}.
}
