% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-summarize.R
\name{cv_aggregate}
\alias{cv_aggregate}
\title{Aggregate CV scores from the results, saved in \code{destin}.}
\usage{
cv_aggregate(
  destin,
  sim = FALSE,
  isim = 1,
  use_meta = TRUE,
  nrep = NULL,
  nfold = NULL,
  cv_gridsize = NULL,
  prob_lambdas = NULL,
  mean_lambdas = NULL
)
}
\arguments{
\item{destin}{Directory with cross-validation output.}

\item{sim}{Simulation or not?}

\item{isim}{Simulation number.}

\item{use_meta}{If TRUE, load \code{nrep}, \code{nfold}, \code{cv_gridsize}, 
\code{prob_lambdas}, and \code{mean_lambdas} from \code{file.path(destin, "meta.Rdata")}.
This is the recommended way to load cross-validation metadata, but we allow the metadata 
to be input directly to this function in case \code{meta.Rdata} was not previously created.}

\item{nrep}{Number of repetitions.}

\item{nfold}{Number of cross-validation folds. Defaults to 5.}

\item{cv_gridsize}{CV grid size.}

\item{prob_lambdas}{Regularization parameters for alphas.}

\item{mean_lambdas}{Regularization parameters for betas.}
}
\description{
Aggregate CV scores from the results, saved in \code{destin}.
}
