% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{det_from_eig}
\alias{det_from_eig}
\title{From eigendecomposition of the sigmas, calculate determinant.}
\usage{
det_from_eig(sigma_eig)
}
\arguments{
\item{sigma_eig}{Eigendecomposition (from \code{get_sigma_eig()}) of sigma.}
}
\value{
Determinant of sigma.
}
\description{
From eigendecomposition of the sigmas, calculate determinant.
}
