% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{draw_membership}
\alias{draw_membership}
\title{Drawing memberships by coinflip using the responsibilities.}
\usage{
draw_membership(resp, verbose = FALSE)
}
\arguments{
\item{resp}{Posteerior cluster probabilities, or responsibilities, from
\code{Estep()}.}
}
\value{
List containing cluster membership draws.
}
\description{
Drawing memberships by coinflip using the responsibilities.
}
