% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-other.R
\name{drawmat_precise_helper}
\alias{drawmat_precise_helper}
\title{Another helper function to /precisely/ draw the entries of a matrix.}
\usage{
drawmat_precise_helper(mat, contour = FALSE, ...)
}
\arguments{
\item{mat}{Matrix of interest.}

\item{contour}{If \code{TRUE}, draw a contour using
\code{lattice::levelplot()}.}

\item{...}{Other arguments to \code{lattice::levelplot()}.}
}
\value{
lattice object.
}
\description{
Another helper function to /precisely/ draw the entries of a matrix.
}
