% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{eigendecomp_sigma}
\alias{eigendecomp_sigma}
\title{(Wrapper for \code{eigendecomp_sigma_barebones()}) Gets eigendecomposition
of a matrix. Basically, \code{sigma == evecs \%*\% Lambdamat \%*\%
solve(evecs)}.}
\usage{
eigendecomp_sigma(sigma)
}
\arguments{
\item{sigma}{A single (dimdat x dimdat) matrix.}
}
\value{
List containing eigenvalues (vector), eigenvectors (matrix of
  eigenvectors as columns)
}
\description{
(Wrapper for \code{eigendecomp_sigma_barebones()}) Gets eigendecomposition
of a matrix. Basically, \code{sigma == evecs \%*\% Lambdamat \%*\%
solve(evecs)}.
}
