% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{eigendecomp_sigma_array}
\alias{eigendecomp_sigma_array}
\title{From a (numclust x dimdat x dimdat) array whose [ii,,]'th entry is
the (dimdat x dimdat) covariance matrix, do eigendecompositions.}
\usage{
eigendecomp_sigma_array(sigma_array)
}
\arguments{
\item{sigma_array}{(numclust x dimdat x dimdat) array.}
}
\value{
|numclust| lengthed list of eigendecompositions.
}
\description{
From a (numclust x dimdat x dimdat) array whose [ii,,]'th entry is
the (dimdat x dimdat) covariance matrix, do eigendecompositions.
}
