% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{eigendecomp_sigma_barebones}
\alias{eigendecomp_sigma_barebones}
\title{Gets eigendecomposition of a matrix. Basically, \code{sigma == evecs \%*\%
Lambdamat \%*\% solve(evecs)}.}
\usage{
eigendecomp_sigma_barebones(sigma)
}
\arguments{
\item{sigma}{dimdat by dimdat matrix}
}
\value{
List containing eigenvalues (vector), eigenvectors (matrix of
  eigenvectors as columns)
}
\description{
Gets eigendecomposition of a matrix. Basically, \code{sigma == evecs \%*\%
Lambdamat \%*\% solve(evecs)}.
}
