% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estep.R
\name{Estep}
\alias{Estep}
\title{Calculates the $k$'th ratio of the (pi * density) of every datapoint,
compared to the sum over all clusters k=1:K. These are called responsibilities (a
posterieri membership probabilities).}
\usage{
Estep(
  mn,
  sigma,
  prob,
  ylist = NULL,
  numclust,
  denslist_by_clust = NULL,
  first_iter = FALSE,
  countslist = NULL
)
}
\arguments{
\item{mn}{Array of all means.}

\item{sigma}{(numclust x dimdat x dimdat) array.}

\item{prob}{Matrix of component weights.}

\item{ylist}{Data.}

\item{numclust}{Number of clusters.}

\item{denslist_by_clust}{Pre-calculated densities.}

\item{first_iter}{\code{TRUE} if this is the first EM iteration, which is
handled separately.}

\item{countslist}{Counts or biomass.}
}
\value{
List of responsibility matrices, containing the posterior
  probabilities of the latent variable $Z$ (memberships to each cluster)
  given the parameter estimate. T-length list of (nt x dimdat)
}
\description{
Calculates the $k$'th ratio of the (pi * density) of every datapoint,
compared to the sum over all clusters k=1:K. These are called responsibilities (a
posterieri membership probabilities).
}
