% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step_sigma_flowcut.R
\name{estep_for_censored_particles}
\alias{estep_for_censored_particles}
\title{Function to assist in calculating Sigma_hat for the censored terms}
\usage{
estep_for_censored_particles(
  iclust,
  resp,
  ylist,
  second_moment_list,
  mn,
  idx_cens_list,
  left_cens_list,
  right_cens_list
)
}
\arguments{
\item{iclust}{cluster index}

\item{resp}{responsibilities list (length TT) of nt x K matrices}

\item{ylist}{list of length K; each [[iclust]] is a list length TT of nt x d matrices (censored y)}

\item{second_moment_list}{list (length K) -> [[iclust]][[tt]][[ii]] = d x d 2nd moment (cond.)}

\item{mn}{fitted means array (TT x d x K)}

\item{cens_mat}{matrix with columns: (tt, ii, d censor flags) where 1 indicates censored}
}
\value{
d x d matrix for censored residual contribution divided by sum of resp for cluster
}
\description{
Function to assist in calculating Sigma_hat for the censored terms
}
