% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estep_flowcut.R
\name{Estep_y_flowcut}
\alias{Estep_y_flowcut}
\title{Computing the new responses conditioned on the observed censoring.}
\usage{
Estep_y_flowcut(
  y,
  X,
  censor_indicator_left,
  censor_indicator_right,
  cens_lim_l_vec,
  cens_lim_u_vec,
  numclust,
  mu_list,
  sigma_list
)
}
\arguments{
\item{y}{response list containing the censored responses. List containing TT elements with each element as a matrix of size ntxdim.}

\item{X}{the matrix of covariates with dimensions tt x d.}

\item{censor_indicator_left}{A list checking whether an observation is left censored. The list if of size TT with each entry being a ntxdim  matrix.}

\item{censor_indicator_right}{A list checking whether an observation is right censored. The list if of size TT with each entry being a ntxdim  matrix.}

\item{cens_lim_l_vec}{Lower censoring limits.}

\item{cens_lim_u_vec}{Upper censoring limits.}

\item{numclust}{number of clusters}

\item{mu_list}{list of means by cluster. Each cluster has a matrix of dim TT x d.}

\item{sigma_list}{list of covariance matrices by cluster. Each cluster has a matrix of dim d x d.}
}
\value{
a nested list containing the responses per time point,observation,cluster in that order of nesting.
}
\description{
Computing the new responses conditioned on the observed censoring.
}
