% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{fill_both_sides}
\alias{fill_both_sides}
\title{Make the remaining region outside of TT white, to prevent spillover of the
cluster means' cex=17 points. (only for 1d data).}
\usage{
fill_both_sides(TT)
}
\description{
Make the remaining region outside of TT white, to prevent spillover of the
cluster means' cex=17 points. (only for 1d data).
}
