% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcut.R
\name{flowcut}
\alias{flowcut}
\title{Main function for our method. Repeats the EM algorithm with |nrep| restarts
(5 by default).}
\usage{
flowcut(..., nrep = 5)
}
\arguments{
\item{...}{Arguments for \code{flowcut_once()}.}

\item{nrep}{Number of restarts.}
}
\value{
The |flowcut| class object that had the best likelihood.
}
\description{
Main function for our method. Repeats the EM algorithm with |nrep| restarts
(5 by default).
}
