% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix-package.R, R/flowmix.R
\name{flowmix}
\alias{flowmix}
\title{Main function for our method. Repeats the EM algorithm with |nrep| restarts
(5 by default).}
\usage{
flowmix(..., nrep = 5)
}
\arguments{
\item{...}{Arguments for \code{flowmix_once()}.}

\item{nrep}{Number of restarts.}
}
\value{
The |flowmix| class object that had the best likelihood.
}
\description{
Main function for our method. Repeats the EM algorithm with |nrep| restarts
(5 by default).
}
