% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix.R
\name{flowmix_once}
\alias{flowmix_once}
\title{Main function for running the EM algorithm once.}
\usage{
flowmix_once(
  ylist,
  X,
  countslist = NULL,
  numclust,
  niter = 1000,
  mn = NULL,
  prob_lambda,
  mean_lambda,
  n_princomps = NULL,
  center = TRUE,
  scale = TRUE,
  n_hidden_nodes = NULL,
  hidden_node_dist = c("unif", "norm"),
  hidden_node_scale = 0.5,
  hidden_node_seed = NULL,
  verbose = FALSE,
  sigma_fac = 1,
  tol_em = 1e-04,
  maxdev = NULL,
  countslist_overwrite = NULL,
  zero_stabilize = FALSE,
  zerothresh = 1e-06,
  admm_rho = 0.01,
  admm_err_rel = 0.001,
  admm_err_abs = 1e-04,
  admm_local_adapt = TRUE,
  admm_local_adapt_niter = 10,
  admm_niter = (if (admm_local_adapt) 1000 else 10000),
  seed = NULL,
  flatX_thresh = 1e-05
)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{X}{Matrix of size (T x p+1)}

\item{countslist}{Multiplicity for particles in \code{ylist}.}

\item{numclust}{Number of clusters}

\item{niter}{Number of EM iterations.}

\item{mn}{Initial means to use. Defaults to NULL.}

\item{prob_lambda}{lambda for lasso for probabilities.}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{n_princomps}{Number of principal components to use if principal components
are to be used as regressors, rather than the original covariates (columns of X).
If \code{NULL}, the original covariates will be used as regressors.}

\item{center}{Whether to center the columns of \code{X} before applying principal components
analysis (PCA).}

\item{scale}{Whether to scale the columns of \code{X} before applying PCA.}

\item{n_hidden_nodes}{Number of hidden nodes to use in the random weight neural network (NN). If
\code{NULL}, no NN is applied as the original covariates are used as regressors.}

\item{hidden_node_dist}{Distribution from which the hidden layer weights are randomly drawn.
Currently supported options are Uniform(-\code{scale}, \code{scale}) (\code{"unif"}) and
Normal(0, \code{scale}).}

\item{hidden_node_scale}{Scale parameter of \code{hidden_node_dist}.}

\item{hidden_node_seed}{RNG seed for drawing hidden layer weights.}

\item{verbose}{TRUE for loudness (e.g. printing EM iterations).}

\item{sigma_fac}{Defaults to 1, and governs how big the initial covariance
matrices should be in size.}

\item{tol_em}{Relative tolerance for EM convergence. Defaults to 1E-4.}

\item{maxdev}{Radius for maximum deviation of cluster means over time.}

\item{countslist_overwrite}{Mainly used by \code{cv.flowmix()}; not to be
modified by user.}

\item{zero_stabilize}{Defaults to FALSE. If TRUE, the EM is only run until
the pattern of zeros in the coefficients stabilizes over EM iterations.}

\item{zerothresh}{Alpha coefficient values below \code{zerothresh} are set
to zero.}

\item{admm_rho}{Step size for ADMM}

\item{admm_err_rel}{Relative error threshold for stopping ADMM.}

\item{admm_err_abs}{Absolute error threshold for stopping ADMM.}

\item{admm_local_adapt}{TRUE if locally adaptive ADMM (LA-ADMM) is to be used. If
so, \code{admm_niter} becomes the inner number of iterations, and
\code{admm_local_adapt_niter} becomes the number of outer iterations.}

\item{admm_local_adapt_niter}{Number of inner iterations in LA ADMM.}

\item{admm_niter}{Number of ADMM iterations.}

\item{seed}{Seven integers that is called directly before \code{init_mn()}
so it can be assigned to \code{.Random.seed} for setting the random state
for the initial mean generation.}

\item{flatX_thresh}{Threshold for detecting if any covariates are flat (low
variance). These flat coefficients will have be set to zero and excluded
from estimation altogether.}
}
\value{
List containing fitted parameters and means and mixture weights,
  across algorithm iterations. \code{beta} is a list of (p+1 x dimdat)
  arrays. \code{alpha} is a (numclust x (p+1)) array.
}
\description{
Main function for running the EM algorithm once.
}
