% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-clusters.R
\name{form_symmetric_kl_distmat}
\alias{form_symmetric_kl_distmat}
\title{From two probability matrices, form a (K x K) distance matrix of the
(n)-vectors. The distance between the vectors is the symmetric KL
divergence.}
\usage{
form_symmetric_kl_distmat(mat1, mat2)
}
\arguments{
\item{mat1}{Matrix 1 of size (n x K).}

\item{mat2}{Matrix 2 of size (n x K).}
}
\value{
K x K matrix containing symmetric KL divergence of each column of
  \code{mat1} and \code{mat2}.
}
\description{
From two probability matrices, form a (K x K) distance matrix of the
(n)-vectors. The distance between the vectors is the symmetric KL
divergence.
}
