% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate.R
\name{gate}
\alias{gate}
\title{Gate (classify) the particles in \code{ylist_particle} according to random
draws from the particle responsibilities -- from the model in \code{res}.}
\usage{
gate(res, ylist_particle, countslist_particle, seed = NULL)
}
\arguments{
\item{res}{flowmix class object.}

\item{ylist_particle}{Particles}

\item{countslist_particle}{Counts or biomass of the particles.}

\item{seed}{Optionally set a seed for drawing particle membership.}
}
\value{
Memberships.
}
\description{
Gate (classify) the particles in \code{ylist_particle} according to random
draws from the particle responsibilities -- from the model in \code{res}.
}
