% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data-flowcut.R
\name{gendat_flowcut}
\alias{gendat_flowcut}
\title{Generic 2d data with 2 clusters}
\usage{
gendat_flowcut(TT = 50, fac = 1, nt = 1000, cens_lim_l_vec, cens_lim_u_vec)
}
\arguments{
\item{TT}{Number of timepoints.}

\item{fac}{Defaults to 1, and determines size of covariance matrix.}

\item{nt}{Number of particles in each time point.}

\item{cens_lim_l_vec}{lower censoring limits; a vector of size dimension.}

\item{cens_lim_u_vec}{upper censoring limits; a vector of size dimension.
points. Defaults to 3/4, in which case the first half of the time points
have cluster relative abundances have a ratio of 6:1:1:0. In the later
half, the fourth cluster appears, and the ratio becomes 6:1:1:1.}
}
\value{
list containing ylist, X with censoting indicators
}
\description{
Generic 2d data with 2 clusters
}
