% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data-flowcut.R
\name{generate_1d_data_flowcut}
\alias{generate_1d_data_flowcut}
\title{Function to create 1d simulated data.}
\usage{
generate_1d_data_flowcut(
  p = 5,
  TT,
  nt = 1000,
  left_limit,
  right_limit,
  prob1 = 3/4,
  sigma1 = NULL,
  sigma2 = NULL
)
}
\arguments{
\item{TT}{Number of time points.}

\item{left_limit}{left censoring limit.}

\item{right_limit}{right censoring limit.}

\item{sigma1}{variance of cluster 1.}

\item{sigma2}{variance of cluster 2.}
}
\value{
list of responses and covariates with cenoring indicators
}
\description{
Function to create 1d simulated data.
}
