% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_data_1d_pseudoreal}
\alias{generate_data_1d_pseudoreal}
\title{Generate pseudoreal, 2-mixture 1d data from the sunlight variable.}
\usage{
generate_data_1d_pseudoreal(
  bin = FALSE,
  seed = NULL,
  datadir = "~/repos/cruisedat/export",
  nt = 1000,
  beta_par = 0.5,
  p = 3,
  dat.gridsize = 30,
  noisetype = c("gaussian", "heavytail", "skewed"),
  df = NULL,
  skew_alpha = NULL,
  gap = 3,
  TT = 100,
  no_changepoint = FALSE,
  gradual_change = FALSE
)
}
\arguments{
\item{bin}{If \code{TRUE}, transform into binned biomass.}

\item{datadir}{Directory that contains the data.}

\item{nt}{Number of points in the first cluster.}

\item{beta_par}{Parameter.}

\item{dat.gridsize}{Grid size for binning.}

\item{noisetype}{Noise type. Default is Gaussian.}

\item{df}{degrees of freedom for t-distribution that replaces N(0,1). This
distribution is further scaled to have mean 0 and variance 1.}

\item{skew_alpha}{skewness of skew-normal distribution that replaces
N(0,1). This distribution is further shifted and scaled to have mean 0 and
variance 1.}

\item{gap}{Size of gap between the two clusters. Defaults to 3.}

\item{no_changepoint}{If TRUE, remove the changepoint covariate
entirely. Defaults to FALSE.}

\item{gradual_change}{If TRUE, the cluster probabilities transition to
low/high from high/low (and vice versa) in the middle. Defaults to FALSE.}

\item{alpha}{Skewness parameter in Azzalini (1985).}
}
\value{
List containing data: {ylist, X, countslist}, and true underlying
  coefficients and mean/probs {mnmat, prob, alpha, beta}.
}
\description{
Generate pseudoreal, 2-mixture 1d data from the sunlight variable.
}
