% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_data_1d_pseudoreal_from_cv}
\alias{generate_data_1d_pseudoreal_from_cv}
\title{Create some artificial but realistic data.}
\usage{
generate_data_1d_pseudoreal_from_cv(
  datadir,
  seed = NULL,
  nt = 100,
  bin = FALSE,
  dat.gridsize = 30
)
}
\arguments{
\item{datadir}{Data directory}

\item{filename}{File name of the "cvres" object from the 2-76-5 experiment}
}
\value{
A list containing the generating coefficients, true means, and data
  (ylist, X, countslist=NULL for now).
}
\description{
Create some artificial but realistic data.
}
