% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_data_generic}
\alias{generate_data_generic}
\title{Generic 2d or 3d data generation.}
\usage{
generate_data_generic(
  p = 5,
  TT = 50,
  fac = 1,
  nt = 1000,
  dimdat = 2,
  seed = NULL,
  prob1 = 3/4
)
}
\value{
list containing ylist, X
}
\description{
Generic 2d or 3d data generation.
}
