% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data-flowcut.R
\name{generate_data_generic_flowcut}
\alias{generate_data_generic_flowcut}
\title{Generic 2d or 3d synthetic data generation.}
\usage{
generate_data_generic_flowcut(
  p = 5,
  TT = 50,
  fac = 1,
  nt = 1000,
  dimdat = 2,
  cens_lim_l_vec,
  cens_lim_u_vec,
  seed = NULL,
  prob1 = 3/4
)
}
\arguments{
\item{p}{Number of covariates.}

\item{TT}{Number of timepoints.}

\item{fac}{Defaults to 1, and determines size of covariance matrix.}

\item{nt}{Number of particles in each time point.}

\item{dimdat}{Dimension of data.}

\item{cens_lim_l_vec}{lower censoring limits; a vector of size dimension.}

\item{cens_lim_u_vec}{upper censoring limits; a vector of size dimension.}

\item{seed}{Random seed.}

\item{prob1}{Relative abundance for cluster 1 in the first half of the time
points. Defaults to 3/4, in which case the first half of the time points
have cluster relative abundances have a ratio of 6:1:1:0. In the later
half, the fourth cluster appears, and the ratio becomes 6:1:1:1.}
}
\value{
list containing ylist, X
}
\description{
Generic 2d or 3d synthetic data generation.
}
