% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-clusters.R
\name{get_best_match_from_kl}
\alias{get_best_match_from_kl}
\title{Compute KL divergence from responsibilities between two models'
responsibilities \code{resp_new} and \code{resp_old}.}
\usage{
get_best_match_from_kl(resp_new, resp_orig)
}
\arguments{
\item{resp_new}{New responsibilities}

\item{resp_orig}{Original responsiblities.}
}
\value{
Calculate reordering \code{o} of the clusters in model represented
  by \code{resp_new}. To be clear, \code{o[i]} of new model is the best
  match with the i'th cluster of the original model.
}
\description{
Compute KL divergence from responsibilities between two models'
responsibilities \code{resp_new} and \code{resp_old}.
}
