% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_frequency}
\alias{get_frequency}
\title{From subsampling bootstrap results (summary files), produce a set of confidence intervals.}
\usage{
get_frequency(nsim, outputdir, origres, ylist_particle, X)
}
\arguments{
\item{outputdir}{Contains files named "summary-(isim).RDS".}

\item{origres}{Model estimated from the entire dataset.}

\item{ylist_particle}{Original particle-level cytogram data.}

\item{X}{Accompanying covariate data.}
}
\description{
From subsampling bootstrap results (summary files), produce a set of confidence intervals.
}
