% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_index_by_clust}
\alias{get_index_by_clust}
\title{Obtain the row numbers of each (binned) cytogram for bootstrap.}
\usage{
get_index_by_clust(drawslist)
}
\arguments{
\item{drawslist}{Membership draws, from \code{draw_membership()}.}
}
\description{
Obtain the row numbers of each (binned) cytogram for bootstrap.
}
