% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lambda-range.R
\name{get_max_lambda}
\alias{get_max_lambda}
\title{A wrapper for \code{calc_max_lambda}. Saves the two maximum lambda values in
a file.}
\usage{
get_max_lambda(
  destin,
  maxres_file = "maxres.Rdata",
  ylist,
  countslist,
  X,
  numclust,
  maxdev,
  max_prob_lambda,
  max_mean_lambda,
  ...
)
}
\arguments{
\item{destin}{Where to save the output (A two-lengthed list called
"maxres").}

\item{maxres_file}{Filename for output. Defaults to maxres.Rdata.}

\item{ylist}{List of responses.}

\item{countslist}{Multiplicity for particles in \code{ylist}.}

\item{X}{Covariates.}

\item{numclust}{Number of clusters.}

\item{maxdev}{Maximum value of lambda.}

\item{max_prob_lambda}{Defaults to 1000.}

\item{max_mean_lambda}{Defaults to 4000.}

\item{...}{Additional arguments to \code{flowmix()}.}
}
\value{
List of the pair of lambda regularization parameter values.
}
\description{
A wrapper for \code{calc_max_lambda}. Saves the two maximum lambda values in
a file.
}
