% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{get_mixture_at_timepoint}
\alias{get_mixture_at_timepoint}
\title{Helper for generating mixture of means}
\usage{
get_mixture_at_timepoint(
  tt,
  nt,
  mnlist,
  pilist,
  sigma = NULL,
  sigmalist = NULL
)
}
\description{
Helper for generating mixture of means
}
