% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{get_range_from_ylist}
\alias{get_range_from_ylist}
\title{Helper: to get ranges for each dimension, of the list of cytograms |ylist|.}
\usage{
get_range_from_ylist(ylist)
}
\arguments{
\item{ylist}{List of cytogram data.}
}
\value{
Matrix whose rows contain the range (min & max) of each dimension.
}
\description{
Helper: to get ranges for each dimension, of the list of cytograms |ylist|.
}
