% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_residuals}
\alias{get_residuals}
\title{Obtain residuals of particles in \code{ylist} from the mean using model
\code{res}, using randomly drawn memberships stored in \code{drawslist}.}
\usage{
get_residuals(ylist, countslist = NULL, res, drawslist, mc.cores = 1)
}
\arguments{
\item{ylist}{Data.}

\item{res}{Model; \code{flowmix} class object.}

\item{drawslist}{Membership draws, from \code{draw_membership()}.}

\item{mc.cores}{Number of cores, to be used by \code{parallel::mclapply()}.}
}
\description{
Obtain residuals of particles in \code{ylist} from the mean using model
\code{res}, using randomly drawn memberships stored in \code{drawslist}.
}
