% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_sigma}
\alias{init_sigma}
\title{Initialize the covariances (naively).}
\usage{
init_sigma(data, numclust, fac = 1)
}
\arguments{
\item{data}{The (nt by 3) datasets. There should be T of them.}

\item{numclust}{Number of clusters.}

\item{fac}{Value to use for the diagonal of the (dimdat x dimdat) covariance
matrix.}
}
\value{
An (K x dimdat x dimdat) array containing the (dimdat by dimdat)
  covariances.
}
\description{
Initialize the covariances (naively).
}
