% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-clusters.R
\name{kl_from_responsibilities}
\alias{kl_from_responsibilities}
\title{Compute KL divergence from responsibilities between two models'
responsibilities \code{resp_new} and \code{resp_old}. Because this can be
computationally intensive, \code{fac=100} allows you to randomly choose only
1/100 of the particles to use for this.}
\usage{
kl_from_responsibilities(resp_new, resp_orig, fac = 100)
}
\arguments{
\item{resp_new}{new responsibilities}

\item{resp_orig}{original responsiblities.}

\item{fac}{Defaults to 100. This means only 1/100 randomly chosen particles
are used for calculating responsibilities}
}
\value{
Calculate KL divergence (of a subset of points)
}
\description{
Compute KL divergence from responsibilities between two models'
responsibilities \code{resp_new} and \code{resp_old}. Because this can be
computationally intensive, \code{fac=100} allows you to randomly choose only
1/100 of the particles to use for this.
}
