% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R
\name{la_admm_oneclust}
\alias{la_admm_oneclust}
\title{LA (locally adaptive) ADMM wrapper to \code{admm_oneclust()}.}
\usage{
la_admm_oneclust(K, ...)
}
\arguments{
\item{K}{Number of outer iterations.}
}
\description{
LA (locally adaptive) ADMM wrapper to \code{admm_oneclust()}.
}
