% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lagpad}
\alias{lagpad}
\title{Helper function to lag a vector}
\usage{
lagpad(x, k)
}
\arguments{
\item{x}{Numeric vector.}

\item{k}{Number of lags.}
}
\value{
Lag-padded numeric vector.
}
\description{
Helper function to lag a vector
}
