% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{loglikelihood_tt_prev}
\alias{loglikelihood_tt_prev}
\title{Second helper function: Calculates one particle's log likelihood *without*
any pre-calculated densities.}
\usage{
loglikelihood_tt_prev(
  ylist,
  tt,
  mu,
  sigma,
  prob,
  countslist = NULL,
  numclust,
  sep = FALSE
)
}
\arguments{
\item{ylist}{Response data.}

\item{tt}{Time point of interest.}

\item{mu}{Means.}

\item{sigma}{Covariances.}

\item{prob}{Population proportions.}

\item{countslist}{(Optional) Counts corresponding to \code{ylist}.}

\item{numclust}{Number of clusters.}

\item{sep}{if TRUE, return separate log likelihood contributions for each particle.}
}
\value{
Log likelihood.
}
\description{
Second helper function: Calculates one particle's log likelihood *without*
any pre-calculated densities.
}
