% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lambda-range.R
\name{logspace}
\alias{logspace}
\title{Helper function to logarithmically space out R.  \code{length} values linear
on the log scale from \code{max} down to \code{min}.}
\usage{
logspace(max, min = NULL, length, min.ratio = 1e-04)
}
\arguments{
\item{max}{Maximum value.}

\item{min}{Minimum value.}

\item{length}{Length of the output string.}

\item{min.ratio}{Factor to multiply to \code{max}.}
}
\value{
Log spaced
}
\description{
Helper function to logarithmically space out R.  \code{length} values linear
on the log scale from \code{max} down to \code{min}.
}
