% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_moments_helper.R
\name{make_condit_moments}
\alias{make_condit_moments}
\title{Helper function to do conditional Gaussian mean, given (1) original
censored particle, and (2) censoring limits and information in each
dimension.}
\usage{
make_condit_moments(
  one_particle,
  mu,
  Sigma,
  left_cens_index,
  right_cens_index,
  cens_lim_l_vec,
  cens_lim_u_vec
)
}
\description{
Helper function to do conditional Gaussian mean, given (1) original
censored particle, and (2) censoring limits and information in each
dimension.
}
