% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-data.R
\name{make_counts}
\alias{make_counts}
\title{Takes a cytogram y that is a (nt x dimdat) matrix, and makes it into a 3d
array that contains the counts for each box.}
\usage{
make_counts(y, grid, qc = NULL)
}
\arguments{
\item{y}{Single cytogram.}

\item{grid}{Grid, created using \code{make_grid()}.}
}
\value{
All counts, as a 3-dimensional array.
}
\description{
Takes a cytogram y that is a (nt x dimdat) matrix, and makes it into a 3d
array that contains the counts for each box.
}
