% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{make_cv_folds_subsample_with_original_membership}
\alias{make_cv_folds_subsample_with_original_membership}
\title{Using subset of indices \code{subsample_inds}, do what \code{make_cv_folds}
does, but using original memberships from data of size \code{orig_TT}. The
new memberships are a subset of those original memberships.}
\usage{
make_cv_folds_subsample_with_original_membership(
  nfold,
  blocksize,
  orig_TT,
  subsample_inds
)
}
\arguments{
\item{nfold}{Number of CV folds.}

\item{blocksize}{Size (number of time indices) of one block.}

\item{orig_TT}{Original data size (for forming original memberships).}

\item{subsample_inds}{Indices, out of \code{1:orig_TT}.}
}
\description{
Using subset of indices \code{subsample_inds}, do what \code{make_cv_folds}
does, but using original memberships from data of size \code{orig_TT}. The
new memberships are a subset of those original memberships.
}
