% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{make_cvscore_filename}
\alias{make_cvscore_filename}
\title{Create file name (a string) for cross-validation results.}
\usage{
make_cvscore_filename(ialpha, ibeta, ifold, irep, sim = FALSE, isim = 1)
}
\description{
Create file name (a string) for cross-validation results.
}
