% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix.R
\name{make_denslist_eigen}
\alias{make_denslist_eigen}
\title{Helper for making list of densities. Returns list by cluster then time
e.g. access by \code{denslist_by_clust[[iclust]][[tt]]}}
\usage{
make_denslist_eigen(
  ylist,
  mu,
  TT,
  dimdat,
  numclust,
  sigma_eig_by_clust,
  countslist
)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x
3), which contains coordinates of the 3-variate particles, organized over
time (T) and with (nt) particles at every time.}

\item{mu}{(T x dimdat x numclust) array.}

\item{TT}{number of time points}

\item{dimdat}{dimension of data.}

\item{numclust}{number of clusters.}

\item{sigma_eig_by_clust}{Result of running
\code{eigendecomp_sigma_array(sigma.list[[iter]])}.}
}
\value{
numclust-lengthed list of TT-lengthed.
}
\description{
Helper for making list of densities. Returns list by cluster then time
e.g. access by \code{denslist_by_clust[[iclust]][[tt]]}
}
