% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_denslist_rcpp}
\alias{make_denslist_rcpp}
\title{Helper for making list of densities. Returns list by cluster then time}
\usage{
make_denslist_rcpp(ylist, mu, TT, dimdat, numclust, sigma_list)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x 3).}

\item{mu}{(T x dimdat x numclust) array, passed as a cube.}

\item{TT}{number of time points (int).}

\item{dimdat}{dimension of data (int).}

\item{numclust}{number of clusters (int).}

\item{sigma_list}{numclust-length list of covariance matrices (dimdat x dimdat).}
}
\value{
numclust-lengthed list of TT-lengthed lists of density vectors.
}
\description{
Helper for making list of densities. Returns list by cluster then time
}
