% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_denslist_rcpp_optimized}
\alias{make_denslist_rcpp_optimized}
\title{Optimized density list maker}
\usage{
make_denslist_rcpp_optimized(ylist, mu, TT, dimdat, numclust, sigma_list)
}
\arguments{
\item{mu}{Armadillo cube (T x dimdat x numclust)}

\item{y_list}{R List of matrices}
}
\description{
Optimized density list maker
}
