% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hidden_nodes.R
\name{make_hidden_nodes}
\alias{make_hidden_nodes}
\title{Create random weight neural network hidden nodes}
\usage{
make_hidden_nodes(X, n.h, a)
}
\arguments{
\item{X}{An n x p design matrix where p is the 
number of covariates (without a column of all 1s 
corresponding to an intercept term).}

\item{n.h}{Number of hidden nodes to output}

\item{a}{Hidden layer weights are drawn from 
Unif(-a, a)}
}
\value{
An n x n.h matrix of hidden layer outputs.
}
\description{
Huang, Zhu, and Siew (2006)
}
