% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{make_iimat_small}
\alias{make_iimat_small}
\title{2d indices for the cross validation jobs.}
\usage{
make_iimat_small(cv_gridsize)
}
\arguments{
\item{cv_gridsize}{CV grid size.}
}
\value{
Integer matrix.
}
\description{
The resulting iimat looks like this:
(#, ialpha, ibeta)
1, 1, 1
2, 1, 2
3, 1, 3
...
225, 15, 15
}
