% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{make_refit_filename}
\alias{make_refit_filename}
\title{Create file name (a string) for re-estimated models for the lambda values
indexed by \code{ialpha} and \code{ibeta}.}
\usage{
make_refit_filename(ialpha, ibeta, irep, sim = FALSE, isim = 1)
}
\description{
Create file name (a string) for re-estimated models for the lambda values
indexed by \code{ialpha} and \code{ibeta}.
}
