% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_seedtab}
\alias{make_seedtab}
\title{Make table of seeds for reproducibility.}
\usage{
make_seedtab(cv_gridsize, nfold, nrep, seeddir, seedfile = "seedtab.csv")
}
\arguments{
\item{cv_gridsize}{CV grid size.}

\item{nfold}{Number of of CV folds.}

\item{nrep}{Number of repetitions.}

\item{seeddir}{Directory in which to save seedtab CSV file.}

\item{seedfile}{Name of seedtab CSV file.}
}
\value{
Tibble of dimension \code{cv_gridsize^2 * (nfold + 1) * nrep} by \code{4 + 7}, 
where the rows correspond to specific calls to \code{flowmix_once}, the first four columns 
correspond to indices of the cross-validation settings (\code{ialpha}, \code{ibeta}, \code{ifold}, 
\code{irep}), and the remaining seven columns are RNG seeds.
}
\description{
Each seed is used to initialize the cluster means in \code{flowmix_once}. If not doing
}
