% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-data.R
\name{make_ybin_3d}
\alias{make_ybin_3d}
\title{Make a (d^3 x 4) matrix of rows that look like (x,y,z,count) from the 3
dimensional (d x d x d) array |counts|.}
\usage{
make_ybin_3d(counts, midpoints, names = NULL)
}
\arguments{
\item{counts}{d x d x d array containing counts. If this is NULL, then dummy
counts of -100 are added.}

\item{midpoints}{midpoints of each bin.}

\item{names}{Names of the data columns.}
}
\value{
(d^3 x 4) matrix of rows that look like (x,y,z,count).
}
\description{
Make a (d^3 x 4) matrix of rows that look like (x,y,z,count) from the 3
dimensional (d x d x d) array |counts|.
}
