% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{manip}
\alias{manip}
\title{Helper to "manipulate" X and y, to get Xtilde and Ytilde and yvec for a more
efficient beta M step (each are |numclust|-length lists, calculated
separately for each cluster).}
\usage{
manip(
  ylist,
  X,
  resp,
  sigma,
  numclust,
  sigma_eig_by_clust = NULL,
  first_iter = FALSE
)
}
\arguments{
\item{sigma}{only used if \code{first_iter} is TRUE. Otherwise,
\code{sigma_eig_by_clust} is used.}

\item{sigma_eig_by_clust}{Eigendecomposition of Sigma.}

\item{first_iter}{TRUE if this is the first EM iteration.}
}
\value{
3 (or dimdat) |numclust|-length lists.
}
\description{
Helper to "manipulate" X and y, to get Xtilde and Ytilde and yvec for a more
efficient beta M step (each are |numclust|-length lists, calculated
separately for each cluster).
}
