% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_moments_helper.R
\name{moment_cal_func_centered}
\alias{moment_cal_func_centered}
\title{Compute the moments of a d-dimensional truncated Gaussian with zero mean with given truncation limits.}
\usage{
moment_cal_func_centered(a_vec, b_vec, Sigma)
}
\arguments{
\item{a_vec}{the lower truncation limits; is a vector of length d.}

\item{b_vec}{the upper truncation limits; is a vector of length d.}

\item{Sigma}{the covariance matrix for the normal of size dxd.}
}
\value{
a list containing the first and second moments.
}
\description{
Compute the moments of a d-dimensional truncated Gaussian with zero mean with given truncation limits.
}
