% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_moments_helper.R
\name{moment_cal_func_centered_1d}
\alias{moment_cal_func_centered_1d}
\title{Compute the moments of a 1-dimensional truncated Gaussian with zero mean with given truncation limits.}
\usage{
moment_cal_func_centered_1d(a_vec, b_vec, Sigma)
}
\arguments{
\item{a_vec}{the lower truncation limit.}

\item{b_vec}{the upper truncation limits.}

\item{Sigma}{the standard deviation.}
}
\value{
a list containing the first and second moments.
}
\description{
Compute the moments of a 1-dimensional truncated Gaussian with zero mean with given truncation limits.
}
