% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R, R/mstep_flowcut.R
\name{Mstep_alpha}
\alias{Mstep_alpha}
\title{Solves the M step for estimating the coefficients for the cluster
probabilities.}
\usage{
Mstep_alpha(resp, X, numclust, lambda, zerothresh = 1e-08)

Mstep_alpha(resp, X, numclust, lambda, zerothresh = 1e-08)
}
\arguments{
\item{resp}{Responsibilities; an (T x nt x K) array.}

\item{X}{Covariate matrix (T x dimdat).}

\item{numclust}{Number of clusters.}

\item{lambda}{Regularization parameter.}

\item{zerothresh}{Values below \code{zerothresh} are set to zero.}
}
\value{
The multinomial logit model coefficients. A matrix of dimension (K x
  (p+1)).

The multinomial logit model coefficients. A matrix of dimension (K x
  (p+1)).
}
\description{
Note, the estimated alphas are not unique, only unique up to a constant
shift.  See \url{https://www.jstatsoft.org/article/view/v033i01} section 4.1
for more details.

Note, the estimated alphas are not unique, only unique up to a constant
shift.  See \url{https://www.jstatsoft.org/article/view/v033i01} section 4.1
for more details.
}
