% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R, R/mstep_flowcut.R
\name{Mstep_beta}
\alias{Mstep_beta}
\title{Solves the M step of beta, using a particular lasso regression
formulation. This is a backup function, and the default beta M-step function
is \code{Mstep_beta_admm()}, which is much faster.}
\usage{
Mstep_beta(
  resp,
  ylist,
  X,
  mean_lambda = 0,
  sigma,
  maxdev = NULL,
  sigma_eig_by_clust = NULL,
  first_iter = FALSE,
  cvxr_ecos_thresh = 1e-08,
  cvxr_scs_eps = 1e-05,
  zerothresh = 1e-08
)

Mstep_beta(
  resp,
  ylist,
  X,
  mean_lambda = 0,
  sigma,
  maxdev = NULL,
  sigma_eig_by_clust = NULL,
  first_iter = FALSE,
  cvxr_ecos_thresh = 1e-08,
  cvxr_scs_eps = 1e-05,
  zerothresh = 1e-08
)
}
\arguments{
\item{resp}{Responsibilities.}

\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{X}{Matrix of size (T x p+1)}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{sigma}{(numclust x dimdat x dimdat) matrix.}

\item{maxdev}{The desired maximum radius of the fitted means from beta0k.}

\item{sigma_eig_by_clust}{Eigendecomposition of Sigma.}

\item{first_iter}{\code{TRUE} if this is the first iteration}

\item{cvxr_ecos_thresh}{CVXR convergence threshold for ECOS solver.}

\item{cvxr_scs_eps}{CVXR convergence threshold for SCS solver.}

\item{zerothresh}{Values below \code{zerothresh} are set to zero.}
}
\value{
Result of M step; a |numclust| length list of (p+1)x(d) matrices,
  each containing the estimated coefficients for the mean estimation.

Result of M step; a |numclust| length list of (p+1)x(d) matrices,
  each containing the estimated coefficients for the mean estimation.
}
\description{
Solves the M step of beta, using a particular lasso regression
formulation. This is a backup function, and the default beta M-step function
is \code{Mstep_beta_admm()}, which is much faster.

Solves the M step of beta, using a particular lasso regression
formulation. This is a backup function, and the default beta M-step function
is \code{Mstep_beta_admm()}, which is much faster.
}
