% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm_flowcut.R
\name{Mstep_beta_admm_flowcut}
\alias{Mstep_beta_admm_flowcut}
\title{The M step of beta, using ADMM. (TODO: This should be able to use the
eigendecomp of the Sigmas for the objective value calculation. That is next
up.)}
\usage{
Mstep_beta_admm_flowcut(
  resp,
  ylist,
  ynew,
  X,
  mean_lambda = 0,
  sigma,
  sigma_eig_by_clust = NULL,
  first_iter = TRUE,
  censor_indicator_left,
  censor_indicator_right,
  cens_lim_l_vec,
  cens_lim_u_vec,
  betas = NULL,
  Zs = NULL,
  Ws = NULL,
  Us = NULL,
  maxdev = NULL,
  niter = 10000,
  rho = 100,
  err_rel = 0.001,
  err_abs = 0,
  zerothresh = 1e-06,
  local_adapt = FALSE,
  local_adapt_niter = 5,
  space = 50
)
}
\arguments{
\item{resp}{Responsibilities.}

\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{X}{Matrix of size (T x p+1)}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{sigma}{(numclust x dimdat x dimdat) matrix.}

\item{sigma_eig_by_clust}{Eigendecomposition of Sigma.}

\item{first_iter}{\code{TRUE} if this is the first iteration}

\item{censor_indicator_left}{list of lower censoring indicators per dimension at each time point tt.}

\item{censor_indicator_right}{list of upper censoring indicators per dimension at each time point tt.}

\item{betas}{Warm-start variable.}

\item{Zs}{Warm-start variable.}

\item{Ws}{Warm-start variable.}

\item{Us}{Warm-start variable.}

\item{maxdev}{The desired maximum radius of the fitted means from beta0k.}

\item{niter}{Total number of outer iterations.}

\item{rho}{todo: fill in}

\item{err_rel}{todo: fill in}

\item{err_abs}{todo: fill in}

\item{zerothresh}{todo: fill in}

\item{local_adapt}{TRUE if locally adaptive ADMM (LA-ADMM) is to be used. If
so, \code{niter} becomes the inner number of iterations, and
\code{local_adapt_niter} becomes the number of outer iterations.}

\item{local_adapt_niter}{Number of outer iterations for LA-ADMM.}

\item{space}{Objectives are only calculated once every \code{space} inner
admm iterations.}
}
\value{
Result of M step; a |numclust| length list of (p+1)x(d) matrices,
  each containing the estimated coefficients for the mean estimation.
}
\description{
The M step of beta, using ADMM. (TODO: This should be able to use the
eigendecomp of the Sigmas for the objective value calculation. That is next
up.)
}
