% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-proximal.R
\name{Mstep_beta_proximal_with_acceleration}
\alias{Mstep_beta_proximal_with_acceleration}
\title{The M step of beta, using proximal gradient descent. Only to be used when
there is no regularization.}
\usage{
Mstep_beta_proximal_with_acceleration(
  resp,
  ylist,
  X,
  mean_lambda = 0,
  sigma,
  sigma_eig_by_clust = NULL,
  first_iter = TRUE,
  betas = NULL,
  maxdev = NULL,
  niter = 10000,
  err_rel = 0.001,
  err_abs = 0,
  space = 50
)
}
\arguments{
\item{resp}{Responsibilities.}

\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{X}{Matrix of size (T x p+1)}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{sigma}{(numclust x dimdat x dimdat) matrix.}

\item{sigma_eig_by_clust}{Eigendecomposition of Sigma.}

\item{first_iter}{\code{TRUE} if this is the first iteration}

\item{betas}{Warm-start variable.}

\item{maxdev}{The desired maximum radius of the fitted means from beta0k.}

\item{niter}{Total number of outer iterations.}

\item{err_rel}{todo: fill in}

\item{err_abs}{todo: fill in}

\item{space}{Objectives are only calculated once every \code{space} inner
admm iterations.}
}
\value{
Result of M step; a |numclust| length list of (p+1)x(d) matrices,
  each containing the estimated coefficients for the mean estimation.
}
\description{
The M step of beta, using proximal gradient descent. Only to be used when
there is no regularization.
}
