% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{Mstep_sigma}
\alias{Mstep_sigma}
\title{M-step for covariance matrix \eqn{\Sigma_k} for cluster \eqn{k} in 1 through
\code{numclust}.}
\usage{
Mstep_sigma(resp, ylist, mn, numclust)
}
\arguments{
\item{resp}{Responsibilities.}

\item{ylist}{List of cytograms.}

\item{mn}{Fitted means.}

\item{numclust}{Number of clusters.}
}
\value{
An array of size (numclust x dimdat x dimdat) containing the
  covariance matrices.
}
\description{
M-step for covariance matrix \eqn{\Sigma_k} for cluster \eqn{k} in 1 through
\code{numclust}.
}
