% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step_sigma_flowcut.R, R/mstep.R,
%   R/mstep_flowcut.R
\name{mtsqrt_inv}
\alias{mtsqrt_inv}
\title{Given a PSD matrix a, compute a^{-1/2}}
\usage{
mtsqrt_inv(a)

mtsqrt_inv(a)

mtsqrt_inv(a)
}
\arguments{
\item{a}{A PSD matrix.}
}
\value{
Matrix of the same size as \code{a}.

Matrix of the same size as \code{a}.
}
\description{
Given a PSD matrix a, compute a^{-1/2}

Given a matrix positive definite matrix a, compute \eqn{a^{-1/2}}.  Only works for
positive semidefinite matrices that are diagonalizable (no normal Jordan
forms, etc.)

Given a matrix positive definite matrix a, compute \eqn{a^{-1/2}}.  Only works for
positive semidefinite matrices that are diagonalizable (no normal Jordan
forms, etc.)
}
