% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{multinom_objective}
\alias{multinom_objective}
\title{Helper function to calculate multinomial objective. This calculates:}
\usage{
multinom_objective(alpha, x, y, lambda, N, exclude.from.penalty = NULL)
}
\arguments{
\item{alpha}{(p x dimdat) matrix of the alpha coefficients.}

\item{x}{(T x p) covariate matrix.}

\item{y}{(T x dimdat) matrix.}

\item{lambda}{regularization parameter value.}

\item{N}{scaling factor for the likelihood function.}

\item{exclude.from.penality}{If not NULL, contains index in 1:p}
}
\value{
Penalized objective value.
}
\description{
\eqn{ \hat \alpha \leftarrow \argmax_{\alpha_{0k}, \alpha_k}
     \frac{1}{N}\sum_{t=1}^T \left( \sum_{k=1}^K \gamma_{\cdot kt}
     (\alpha_{0k} + {X^{(t)}}^T \alpha_k) - n_t \log \sum_{l=1}^K
     \exp(\alpha_{0l} + {X^{(t)}}^T \alpha_l) \right) - \lambda_\alpha
     \sum_{k=1}^K \|\alpha_k\|_1}
}
