% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcut.R
\name{my_convert}
\alias{my_convert}
\title{Convert the old format to the new.}
\usage{
my_convert(ylist, cens_indicator_list_left, cens_indicator_list_right)
}
\arguments{
\item{ylist}{original ylist}

\item{cens_indicator_list_left}{original left censor indicator}

\item{cens_indicator_list_right}{original right censor indicator.}
}
\value{
Separate list of responses, uncensored and censored. Also, two list
  objects that mark, in which (of the dimdat) dimension the censoring has
  happened, left or right.
}
\description{
Convert the old format to the new.
}
