% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_moments_helper.R
\name{my_pmvnorm}
\alias{my_pmvnorm}
\title{Computes the normal probabilities giving a small lower bound to very low probability events}
\usage{
my_pmvnorm(lower, upper, mean, sigma)
}
\arguments{
\item{lower}{d-dimensional vector of lower entries. Can be thought of as the lowest edge in the cube considered.}

\item{upper}{d-dimensional vector of upper entries. Can be thought of as the highest edge in the cube considered.}

\item{mean}{the d-dimensional vector of means for the multivariate normal.}

\item{sigma}{the dxd covariance matrix for the multivariate normal.}
}
\value{
a probability value. This value is lower bounded by 1e-12.
}
\description{
Computes the normal probabilities giving a small lower bound to very low probability events
}
