% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcut.R
\name{my_reassemble}
\alias{my_reassemble}
\title{An opposite operation from the separation in myconvert().}
\usage{
my_reassemble(ylist_uncens, ylist_cens, idx_cens_list, idx_uncens_list)
}
\description{
An opposite operation from the separation in myconvert().
}
