% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R
\name{myschur}
\alias{myschur}
\title{Convenience function for obtaining a Schur decomposition of a matrix
\code{mat}. Used for simplifying a Sylvester equation.}
\usage{
myschur(mat)
}
\arguments{
\item{mat}{Numeric matrix.}
}
\value{
List.
}
\description{
Convenience function for obtaining a Schur decomposition of a matrix
\code{mat}. Used for simplifying a Sylvester equation.
}
