% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Objectives (negative log likelihood).}
\usage{
objective(
  mu,
  prob,
  sigma,
  ylist,
  prob_lambda = 0,
  mean_lambda = 0,
  alpha = NULL,
  beta = NULL,
  denslist_by_clust = NULL,
  countslist = NULL,
  each = FALSE,
  sep = FALSE
)
}
\arguments{
\item{mu}{array of dimension (T x dimdat x numclust ) ; old: T by M by p.}

\item{prob}{matrix of mixture proportions, T by M.}

\item{sigma}{array of dimension T by M by p by p.}

\item{alpha}{linear coefficients for regression on (log ratio of) prob.}

\item{beta}{linear coefficients for regression on mean.}

\item{data}{TT lengthed list of data}

\item{N}{Sum of the counts/biomass across all cytograms.}
}
\value{
Either the original objective value or, T-lengthed vector of
  objective values at each time point.
}
\description{
Objectives (negative log likelihood).
}
