% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective_newdat}
\alias{objective_newdat}
\title{A convenience function to calculate the CV score, or out-of-sample objective
function (no penalization), given new data.}
\usage{
objective_newdat(res, ylist, countslist)
}
\arguments{
\item{res}{flowmix object.}

\item{ylist}{New data.}

\item{countslist}{New count data.}
}
\value{
An out-of-sample, unpenalized likelihood (i.e. CV score).
}
\description{
A convenience function to calculate the CV score, or out-of-sample objective
function (no penalization), given new data.
}
