% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{one_job}
\alias{one_job}
\title{Helper function to run ONE job for CV, in ialpha, ibeta, ifold, irep.}
\usage{
one_job(
  ialpha,
  ibeta,
  ifold,
  irep,
  folds,
  destin,
  mean_lambdas,
  prob_lambdas,
  sim = FALSE,
  isim = 1,
  seedtab = NULL,
  ylist,
  countslist,
  X,
  ...
)
}
\arguments{
\item{ialpha}{Index for alpha.}

\item{ibeta}{Index for beta.}

\item{ifold}{Index for CV folds.}

\item{irep}{Index for 1 through nrep.}

\item{folds}{CV folds.}

\item{destin}{Destination directory.}

\item{mean_lambdas}{List of regularization parameters for mean model.}

\item{prob_lambdas}{List of regularization parameters for prob model.}

\item{sim}{Set to \code{TRUE} if this is a simulation; the file name
containing results for, say `isim=3`, has "3-" appended to the beginning.}

\item{isim}{Simulation number.}

\item{ylist}{Data.}

\item{countslist}{Counts or biomass.}

\item{X}{Covariates.}

\item{...}{Rest of arguments for \code{flowmix_once()}.}
}
\value{
Nothing is returned. Instead, a file named "1-1-1-1-cvscore.Rdata"
  is saved in \code{destin}. (The indices here are ialpha-ibeta-ifold-irep).

If this is a simulation, the file name containing results for, say `isim=3`,
  has "3-" appended to the beginning.
}
\description{
Helper function to run ONE job for CV, in ialpha, ibeta, ifold, irep.
}
