% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{one_job_refit_flowcut}
\alias{one_job_refit_flowcut}
\title{Refit model for one pair of regularization parameter values. Saves to
\code{nrep} files named like "1-4-3-fit.Rdata", for
"(ialpha)-(ibeta)-(irep)-fit.Rdata".}
\usage{
one_job_refit_flowcut(
  ialpha,
  ibeta,
  destin,
  mean_lambdas,
  prob_lambdas,
  seedtab = NULL,
  ylist,
  countslist,
  X,
  cens_ind_left,
  cens_ind_right,
  cens_vec_1,
  cens_vec_2,
  sim = FALSE,
  isim = 1,
  ...
)
}
\arguments{
\item{ialpha}{Index for alpha.}

\item{ibeta}{Index for beta.}

\item{destin}{Destination directory.}

\item{mean_lambdas}{List of regularization parameters for mean model.}

\item{prob_lambdas}{List of regularization parameters for prob model.}

\item{seedtab}{A table containing seeds (7 columns)}

\item{ylist}{Data.}

\item{countslist}{Counts or biomass.}

\item{X}{Covariates.}

\item{cens_ind_left}{the left censoring indicators for the data.}

\item{cens_ind_right}{the right censoring indicators for the data.}

\item{cens_vec_1}{the lower censoring values.}

\item{cens_vec_2}{the upper censoring values.}

\item{sim}{Set to \code{TRUE} if this is a simulation; the file name
containing results for, say `isim=3`, has "3-" appended to the beginning.}

\item{isim}{Simulation number.}

\item{...}{Rest of arguments for \code{flowmix_once()}.}
}
\description{
(Note, \code{nrep} is not an input to this function.)
}
