% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R, R/mstep-admm_flowcut.R
\name{outer_converge}
\alias{outer_converge}
\title{Check if 4 consecutive objective values are sufficiently close to 1.}
\usage{
outer_converge(objectives)

outer_converge(objectives)
}
\arguments{
\item{objectives}{Numeric vector.}
}
\value{
Boolean.

Boolean.
}
\description{
Check if 4 consecutive objective values are sufficiently close to 1.

Check if 4 consecutive objective values are sufficiently close to 1.
}
