% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d-extra.R
\name{plot3d_compare.covarem}
\alias{plot3d_compare.covarem}
\title{Temporary function for plotting three different 3d models together. Useful
but not included in the package.}
\usage{
plot3d_compare.covarem(
  obj1,
  obj2,
  obj3,
  ylist,
  countslist = NULL,
  tt,
  cex.fac = 1,
  destin = NULL
)
}
\description{
Temporary function for plotting three different 3d models together. Useful
but not included in the package.
}
