% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette-plot-helpers.R
\name{plot3d_one_3d_plot}
\alias{plot3d_one_3d_plot}
\title{Making a 3d scatter plot with a certain angle..}
\usage{
plot3d_one_3d_plot(
  ylist,
  obj = NULL,
  tt,
  countslist = NULL,
  phi = 40,
  cex.fac = 1,
  cex.axis = 1,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  ticktype = "detailed",
  col = rgb(0, 0, 1, 0.1),
  ...
)
}
\arguments{
\item{...}{Additional arguments to plot3d::scatter3d().}
}
\description{
Making a 3d scatter plot with a certain angle..
}
\keyword{internal}
