% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-3d.R
\name{plot3d_simple}
\alias{plot3d_simple}
\title{A simple function for plotting 3-dimensional data.}
\usage{
plot3d_simple(
  obj,
  ylist,
  countslist = NULL,
  tt,
  cex.fac.2d = 1,
  pt_col = rgb(0, 0, 1, 0.1),
  cex.fac.3d = 1
)
}
\description{
A simple function for plotting 3-dimensional data.
}
