% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{plot_1d}
\alias{plot_1d}
\title{Plot 1 dimensional model results.}
\usage{
plot_1d(
  ylist,
  countslist,
  res = NULL,
  scale = TRUE,
  main = "",
  time_axis = FALSE,
  time = NULL,
  mn.scale = 5,
  cex_clust_label = 1.5,
  omit_band = FALSE,
  omit_label = FALSE,
  reorder_clusters = TRUE,
  cex_data = 1,
  ylim = NULL,
  cols = NULL
)
}
\arguments{
\item{ylist}{Data.}

\item{countslist}{Multiplicity.}

\item{res}{Optionally, |covarem| object of fitted model.}

\item{scale}{Defaults to TRUE. If TRUE, scale the data colors, relative to
the largest count in |countslist|.}

\item{main}{Plot title.}

\item{time}{Vector of date strings.}
}
\value{
NULL.
}
\description{
Plot 1 dimensional model results.
}
