% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2d.R
\name{plot_2d}
\alias{plot_2d}
\title{For 2-dimensional data, make a single scatterplot for one time points' worth
of data.}
\usage{
plot_2d(
  ylist,
  tt,
  res = NULL,
  drawslist = NULL,
  resp = NULL,
  iclust = NULL,
  cols = NULL
)
}
\arguments{
\item{ylist}{Data.}

\item{tt}{Time point.}

\item{res}{flowmix object.}

\item{drawslist}{A list containing the randomly generated memberships using
the posterior cluster probabilities, which can be generated using
\code{draw_membership()}. Defaults to NULL.}

\item{resp}{Posterior probabilities of cluster memberships
(responsibilities). Defaults to NULL.}

\item{iclust}{Cluster number(s). Defaults to NULL.}

\item{cols}{Colors for each cluster. Defaults to NULL.}
}
\description{
CAN'T handle the binned case (with \code{countslist}) yet.
}
