% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-3d.R
\name{plot_3d}
\alias{plot_3d}
\title{Main 3d plotting function. It's possible to call this function without an
object |obj|. IN that case, it is a pure data plotting function.}
\usage{
plot_3d(
  obj,
  ylist,
  countslist = NULL,
  tt,
  show.xb.constraint = FALSE,
  cex.fac.2d = 1,
  par_cex_2d = 1,
  pt_col = rgb(0, 0, 1, 0.1),
  cex.fac.3d = 1,
  lon = NULL,
  lat = NULL
)
}
\arguments{
\item{obj}{flowmix object.}

\item{ylist}{List of covariates.}

\item{countslist}{List of counts. Defaults to null.}

\item{tt}{time point of interest.}

\item{show.xb.constraint}{If TRUE, show the ball constraint boundaries.}
}
\description{
*Don't* do this.
}
