% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-other.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plots the time series of covariates from an |obj| object.}
\usage{
plot_covariates(obj, tt = NULL)
}
\arguments{
\item{obj}{An object of class |flowmix|.}

\item{tt}{Time point to highlight.}
}
\value{
no return
}
\description{
Plots the time series of covariates from an |obj| object.
}
