% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette-helpers.R
\name{plot_cvscore}
\alias{plot_cvscore}
\title{For a matrix of CV scores (which are included in output from the function
\code{aggregateres()} or \code{blockcv_summary()}, make a 2d heatmap}
\usage{
plot_cvscore(cvscore.mat)
}
\arguments{
\item{cvscore.mat}{Matrix containing CV scores.}
}
\description{
For a matrix of CV scores (which are included in output from the function
\code{aggregateres()} or \code{blockcv_summary()}, make a 2d heatmap
}
\keyword{internal}
