% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-other.R
\name{plot_prob}
\alias{plot_prob}
\title{Plot cluster probabilities over time.}
\usage{
plot_prob(res, iclusts = NULL, main = NULL, cols = NULL)
}
\arguments{
\item{res}{flowmix object.}

\item{iclusts}{Optionally, provide the cluster numbers to plot only a
subset.}
}
\description{
Plot cluster probabilities over time.
}
