% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{plot_ylist}
\alias{plot_ylist}
\title{Make 1d data plot. No axes are added.}
\usage{
plot_ylist(
  ylist,
  countslist,
  res = NULL,
  scale = TRUE,
  main = "",
  cex = 1,
  ylim = NULL
)
}
\arguments{
\item{ylist}{Data.}

\item{countslist}{Multiplicity.}

\item{res}{Optionally, |covarem| object of fitted model.}

\item{scale}{Defaults to TRUE. If TRUE, scale the data colors, relative to
the largest count in |countslist|.}

\item{main}{Plot title.}
}
\description{
Make 1d data plot. No axes are added.
}
