% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.flowmix}
\alias{print.flowmix}
\title{Default printing functionality for flowmix() results.}
\usage{
\method{print}{flowmix}(obj)
}
\arguments{
\item{obj}{An object of class "flowmix"}
}
\value{
Nothing.
}
\description{
Default printing functionality for flowmix() results.
}
