% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_progress}
\alias{print_progress}
\title{A helper function to print the progress of a loop or simulation.}
\usage{
print_progress(
  isim,
  nsim,
  type = "simulation",
  lapsetime = NULL,
  lapsetimeunit = "seconds",
  start.time = NULL,
  fill = FALSE
)
}
\arguments{
\item{isim}{Replicate number.}

\item{nsim}{Total number of replicates.}

\item{type}{Type of job you're running. Defaults to "simulation".}

\item{lapsetime}{Lapsed time, in seconds (by default).}

\item{lapsetimeunit}{"second".}

\item{start.time}{start time.}

\item{fill}{Whether or not to fill the line.}
}
\value{
No return
}
\description{
A helper function to print the progress of a loop or simulation.
}
