% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reformat_coef}
\alias{reformat_coef}
\title{Prettifying the coefficients alpha and beta.}
\usage{
reformat_coef(alpha, beta, numclust, dimdat, X)
}
\arguments{
\item{alpha}{Alpha coefficients.}

\item{beta}{Beta coefficients.}

\item{numclust}{Number of clusters}

\item{dimdat}{dimension of data.}

\item{X}{Covariates.}
}
\value{
list containing prettified alpha and beta.
}
\description{
Prettifying the coefficients alpha and beta.
}
