% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-clusters.R
\name{reorder_clust}
\alias{reorder_clust}
\title{Reorder the results of one flowmix object so that cluster 1 through
\code{numclust} is in a particular order. The default is decreasing order of
the averages (over time) of the cluster means.}
\usage{
reorder_clust(res, ord = NULL)
}
\arguments{
\item{res}{flowmix object}

\item{ord}{Defaults to NULL. Use if you have an ordering in mind.}
}
\value{
Same object, but with clusters reordered.
}
\description{
Reorder the results of one flowmix object so that cluster 1 through
\code{numclust} is in a particular order. The default is decreasing order of
the averages (over time) of the cluster means.
}
