% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-3d.R
\name{scatterplot_2d}
\alias{scatterplot_2d}
\title{Making data plot for two dimensions of the original three (those in |ind|).}
\usage{
scatterplot_2d(
  ylist,
  obj,
  tt,
  countslist = NULL,
  dims = c(1, 2),
  cex_fac = 10,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  constant_total_count = FALSE,
  add_clust_labels = TRUE,
  only_model = FALSE,
  subset_of_clust = NULL,
  col = NULL,
  lty = 2,
  lwd = 1/2,
  not_so_small_cex = FALSE,
  mn_cex_fac = 5,
  pt_col = rgb(0, 0, 1, 0.1),
  par_cex = 1
)
}
\description{
Making data plot for two dimensions of the original three (those in |ind|).
}
