% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-3d.R
\name{scatterplot_2d_addmodel}
\alias{scatterplot_2d_addmodel}
\title{Helper function for \code{scatterplot_2d()}, to add means and ellipses
for confidence regions.}
\usage{
scatterplot_2d_addmodel(
  obj,
  tt,
  dims,
  col = "tomato",
  add_clust_labels = TRUE,
  clust_labels = NULL,
  ellipse = TRUE,
  subset_of_clust = NULL,
  lwd = 1/2,
  lty = 2,
  not_so_small_cex = FALSE,
  mn_cex_fac = 5
)
}
\arguments{
\item{obj}{A |covarem| class object.}

\item{tt}{time point.}

\item{dims}{e.g. \code{c(1,2)}, containing the two dimensions you want.}

\item{col}{color of the added means and confidence regions.}
}
\value{
NULL.
}
\description{
Helper function for \code{scatterplot_2d()}, to add means and ellipses
for confidence regions.
}
