% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{sigma_half_from_eig}
\alias{sigma_half_from_eig}
\title{From the eigendecomposition, make sigma halves.}
\usage{
sigma_half_from_eig(sigma_eig)
}
\arguments{
\item{sigma_eig}{Eigendecomposition (from \code{get_sigma_eig()}) of sigma.}
}
\value{
Square root of sigma i.e. if the returned object is
  \code{sigma_half}, we should have \code{sigma_half %*% sigma_half ==
  sigma}.
}
\description{
From the eigendecomposition, make sigma halves.
}
