% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step_sigma_flowcut.R
\name{Sigma_hat_calculation_fn}
\alias{Sigma_hat_calculation_fn}
\title{Function to assist in calculating Sigma_hat for the censored terms}
\usage{
Sigma_hat_calculation_fn(iclust, resp, ylist, second_moment_list, mn, cens_mat)
}
\arguments{
\item{iclust}{cluster index}

\item{resp}{responsibilities list (length TT) of nt x K matrices}

\item{ylist}{list of length K; each [[iclust]] is a list length TT of nt x d matrices (censored y)}

\item{second_moment_list}{list (length K) -> [[iclust]][[tt]][[ii]] = d x d 2nd moment (cond.)}

\item{mn}{fitted means array (TT x d x K)}

\item{cens_mat}{matrix with columns: (tt, ii, d censor flags) where 1 indicates censored}
}
\value{
d x d matrix for censored residual contribution divided by sum of resp for cluster
}
\description{
Function to assist in calculating Sigma_hat for the censored terms
}
