% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{sigma_inv_from_eig}
\alias{sigma_inv_from_eig}
\title{From the eigendecomposition, make inverse of sigma.}
\usage{
sigma_inv_from_eig(sigma_eig)
}
\arguments{
\item{sigma_eig}{Eigendecomposition (from \code{get_sigma_eig()}) of sigma.}
}
\value{
Inverse of sigma.
}
\description{
From the eigendecomposition, make inverse of sigma.
}
