% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R
\name{soft_thresh}
\alias{soft_thresh}
\title{(Helper) Soft thresholding of |a| at radius of |b|.}
\usage{
soft_thresh(a, b)
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector of same length as \code{length(b)}.}
}
\value{
Soft-thresholded version of a and b.
}
\description{
(Helper) Soft thresholding of |a| at radius of |b|.
}
