% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sym_diff}
\alias{sym_diff}
\title{Symmetric difference.}
\usage{
sym_diff(a, b)
}
\arguments{
\item{a}{One vector.}

\item{b}{Another vector.}
}
\value{
Symmetric difference of discrete set \code{a} and \code{b}.
}
\description{
Symmetric difference.
}
