% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-clusters.R
\name{symmetric_kl}
\alias{symmetric_kl}
\title{Symmetric KL divergence, of two probability vectors.}
\usage{
symmetric_kl(vec1, vec2)
}
\arguments{
\item{vec1}{First probability vector.}

\item{vec2}{Second prbability vector.}
}
\value{
Symmetric KL divergence (scalar).
}
\description{
Symmetric KL divergence, of two probability vectors.
}
