% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim}
\alias{trim}
\title{Trim data so that both the list of binned responses and counts don't have
any zeros.}
\usage{
trim(ybin_list, counts_list)
}
\arguments{
\item{ybin_list}{List of binned data, assumed to be \code{d^3} lengthed.}

\item{counts_list}{Counts of binned data, assumed to be \code{d^3} lengthed.}
}
\value{
Trimmed data, each of different length.
}
\description{
Trim data so that both the list of binned responses and counts don't have
any zeros.
}
