% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-data.R
\name{trim_one_cytogram}
\alias{trim_one_cytogram}
\title{Trim single cytogram so that both the list of binned responses and counts
don't have any zeros.}
\usage{
trim_one_cytogram(ybin, counts)
}
\arguments{
\item{ybin}{Binned data (assumed to be d^3 lengthed)}

\item{counts}{Counts of binned data (assumed to be d^3 lengthed)}
}
\description{
Trim single cytogram so that both the list of binned responses and counts
don't have any zeros.
}
