% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{add_lagpar}
\alias{add_lagpar}
\title{(To add to package) add step function to the lags}
\usage{
add_lagpar(X, lags)
}
\arguments{
\item{X}{Covariate matrix.}

\item{lags}{Number of lags.}
}
\value{
Transformed X.

Helper function to lag a vector
}
\description{
(To add to package) add step function to the lags
}
