% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{add_transition}
\alias{add_transition}
\title{(to add to package) Add step function bases demarcated by transition region
crossings.}
\usage{
add_transition(X, lat)
}
\description{
(to add to package) Add step function bases demarcated by transition region
crossings.
}
