% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alter_alpha}
\alias{alter_alpha}
\title{Altering the beta coefficient (for all clusters) by adding zero coefficients
to certain rows.}
\usage{
alter_alpha(alpha, flatX, orig_names)
}
\arguments{
\item{alpha}{alpha coefficients (One numclust x p matrix).}

\item{flatX}{Indices, out of 1 through p, of which variables are flat.}

\item{orig_names}{Original column names of X.}
}
\value{
Same format as alpha, but with zeros in the columns where variable
  that was flat.
}
\description{
Altering the beta coefficient (for all clusters) by adding zero coefficients
to certain rows.
}
