% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_moments_helper.R
\name{cens_cond_normal}
\alias{cens_cond_normal}
\title{Compute the moments of a d-dimensional truncated Gaussian with zero mean with given truncation limits.}
\usage{
cens_cond_normal(
  ii,
  tt,
  y,
  X,
  censor_indicator_left,
  censor_indicator_right,
  cens_lim_l_vec,
  cens_lim_u_vec,
  numclust,
  mu_list,
  sigma_list
)

cens_cond_normal(
  ii,
  tt,
  y,
  X,
  censor_indicator_left,
  censor_indicator_right,
  cens_lim_l_vec,
  cens_lim_u_vec,
  numclust,
  mu_list,
  sigma_list
)
}
\arguments{
\item{ii}{the particle index for which the censored}

\item{tt}{the time point in question}

\item{y}{the response. Is a list with a matrix at each time point of dimension nt x dimension.}

\item{X}{the matrix of covariates with dimensions tt x d.}

\item{censor_indicator_left}{A list checking whether an observation is left censored. The list if of size TT with each entry being a ntxdim  matrix.}

\item{censor_indicator_right}{A list checking whether an observation is right censored. The list if of size TT with each entry being a ntxdim  matrix.}

\item{cens_lim_l_vec}{Lower censoring limits.}

\item{cens_lim_u_vec}{Upper censoring limits.}

\item{numclust}{number of clusters}

\item{mu_list}{list of means by cluster. Each cluster has a matrix of dim TT x d.}

\item{sigma_list}{list of covariance matrices by cluster. Each cluster has a matrix of dim d x d.}
}
\value{
a list for the means and variances of the censored/uncensored observation ii at time tt per each cluster.

a list for the means and variances of the censored/uncensored observation ii at time tt per each cluster.
}
\description{
Compute the moments of a d-dimensional truncated Gaussian with zero mean with given truncation limits.

Compute the moments of a d-dimensional truncated Gaussian with zero mean with given truncation limits.
}
