% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_trim}
\alias{check_trim}
\title{Helper to check whether ylist and countslist have been trimmed of zero,
e.g. using \code{trim()}.}
\usage{
check_trim(ylist, countslist)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{countslist}{Multiplicity for particles in \code{ylist}.}
}
\description{
Helper to check whether ylist and countslist have been trimmed of zero,
e.g. using \code{trim()}.
}
