% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_moments_helper.R
\name{cond_mean_var_func}
\alias{cond_mean_var_func}
\title{Compute conditional (lower-dimensional) multivariate normal (TODO: if needed, y can turned into a matrix).}
\usage{
cond_mean_var_func(y, mu, Sigma, observed_dims)
}
\arguments{
\item{y}{d-dimensional observation that may be censored in some dimensions.}

\item{mu}{d-dimensionl Normal mean.}

\item{Sigma}{Normal dxd covariance matrix.}

\item{observed_dims}{A subset of 1,2,..,d on which we condition}
}
\value{
list containing mean and variance of the conditional multivariate
  normal.
}
\description{
Compute conditional (lower-dimensional) multivariate normal (TODO: if needed, y can turned into a matrix).
}
