% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-summarize.R
\name{cv_aggregate}
\alias{cv_aggregate}
\title{Aggregate CV scores from the results, saved in \code{destin}.}
\usage{
cv_aggregate(destin, sim = FALSE, isim = 1)
}
\arguments{
\item{destin}{Directory with cross-validation output.}

\item{sim}{Simulation or not?}

\item{isim}{Simulation number.}
}
\description{
Aggregate CV scores from the results, saved in \code{destin}.
}
