% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-summarize.R
\name{cv_aggregate_res}
\alias{cv_aggregate_res}
\title{Helper to aggregate parallelized CV results and obtain the |res| object, all
saved in |destin|.}
\usage{
cv_aggregate_res(destin, sim = FALSE, isim = NULL)
}
\arguments{
\item{destin}{Directory with cross-validation output.}

\item{sim}{Simulation or not?}

\item{isim}{Simulation number.}
}
\value{
List containing, for every (ialpha, ibeta), the "best" estimated
  model out of the |nrep| replicates (best in the sense that it had the best
  likelihood value out of the |nrep| replicates.)
}
\description{
Helper to aggregate parallelized CV results and obtain the |res| object, all
saved in |destin|.
}
