% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-summarize.R
\name{cv_summary}
\alias{cv_summary}
\title{Main function for summarizing the cross-validation results.}
\usage{
cv_summary(destin = ".", save = FALSE, filename = "summary.RDS")
}
\arguments{
\item{destin}{Where to save the output.}

\item{save}{If TRUE, save to \code{file.path(destin, filename)}.}

\item{filename}{File name to save to.}
}
\value{
List containing various outcomes from the cross-validation, such as
  \code{bestres} which is the \code{flowmix} class object of the overall
  best model chosen from the cross-validation; \code{cvscoremat} containing
  a 2d matrix of CV scores from all pairs of lambdas; \code{bestreslist}
  contains all the best models (out of \code{nrep} EM replications)q from the
  each pair of lambda values. If \code{isTRUE(save)}, nothing is returned.
}
\description{
Main function for summarizing the cross-validation results.
}
