% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix-package.R
\docType{package}
\name{flowmix-package}
\alias{flowmix}
\alias{flowmix-package}
\title{flowmix: computing sparse mixture of regressions for flow cytometry data.}
\description{
The main function for conducting analyses is \code{flowmix()}.
}
\details{
Try the built-in vignette (`vignette("flowmix")`), for a step-by-step
tutorial.
}
\author{
\strong{Maintainer}: Sangwon Hyun \email{robohyun66@gmail.com}

}
\keyword{internal}
