% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix-package.R, R/flowmix.R
\docType{package}
\name{flowmix}
\alias{flowmix}
\title{flowmix: computing sparse mixture of regressions for flow cytometry data.}
\usage{
flowmix(..., nrep = 5)
}
\arguments{
\item{...}{Arguments for \code{flowmix_once()}.}

\item{nrep}{Number of restarts.}
}
\value{
The |flowmix| class object that had the best likelihood.
}
\description{
The main function for conducting analyses is \code{flowmix()}.
}
\details{
Try the built-in vignette (`vignette("flowmix")`), for a step-by-step
tutorial.
}
