% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix.R
\name{flowmix_once}
\alias{flowmix_once}
\title{Main function for running the EM algorithm once.}
\usage{
flowmix_once(
  ylist,
  X,
  countslist = NULL,
  numclust,
  niter = 1000,
  mn = NULL,
  prob_lambda,
  mean_lambda,
  verbose = FALSE,
  sigma_fac = 1,
  tol_em = 1e-04,
  maxdev = NULL,
  countslist_overwrite = NULL,
  zero_stabilize = FALSE,
  init_mn_flatten = FALSE,
  mstep_cvxr_ecos_thresh = 1e-08,
  mstep_cvxr_scs_eps = 1e-05,
  zerothresh = 1e-06,
  admm = TRUE,
  admm_rho = 0.01,
  admm_err_rel = 0.001,
  admm_err_abs = 1e-04,
  admm_local_adapt = TRUE,
  admm_local_adapt_niter = 10,
  admm_niter = (if (admm_local_adapt) 1000 else 10000),
  CVXR = FALSE
)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x
3), which contains coordinates of the 3-variate particles, organized over
time (T) and with (nt) particles at every time.}

\item{X}{Matrix of size (T x p+1)}

\item{countslist}{Multiplicity for particles in \code{ylist}.}

\item{numclust}{Number of clusters}

\item{prob_lambda}{lambda for lasso for probabilities.}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{tol_em}{Relative tolerance for EM convergence. Defaults to 1E-4.}

\item{zero_stabilize}{Defaults to FALSE. If TRUE, the EM is only run until
the zero pattern in the coefficients stabilize.}

\item{prob.list}{(T by K)}
}
\value{
List containing fitted parameters and means and mixture weights,
  across algorithm iterations. \code{beta} is a list of (p+1 x dimdat)
  arrays. \code{alpha} is a (numclust x (p+1)) array.
}
\description{
Main function for running the EM algorithm once.
}
