% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_1d_data}
\alias{generate_1d_data}
\title{Function to create 1d simulated data.}
\usage{
generate_1d_data(TT, sigma1 = NULL, sigma2 = NULL)
}
\arguments{
\item{TT}{Number of time points.}

\item{sigma1}{variance of cluster 1.}

\item{sigma2}{variance of cluster 2.}
}
\description{
Function to create 1d simulated data.
}
