% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_1d_data}
\alias{generate_1d_data}
\title{Function to create 1d simulated data.}
\usage{
generate_1d_data(TT, sigma1 = NULL, sigma2 = NULL)
}
\description{
Function to create 1d simulated data.
}
