% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_frequency}
\alias{get_frequency}
\title{From subsampling bootstrap results (summary files), produce stability
measures.}
\usage{
get_frequency(
  nsim,
  outputdir,
  origres,
  ylist_particle,
  X,
  prefix = "summary-sim-"
)
}
\arguments{
\item{outputdir}{Contains files named "summary-(isim).RDS".}

\item{origres}{Model estimated from the entire dataset.}

\item{ylist_particle}{Original particle-level cytogram data.}

\item{X}{Accompanying covariate data.}

\item{prefix}{The prefix of the summary files, e.g. "summary-sim-" for file that are named "summary-sim-95.RDS".}
}
\value{
List of nonzero frequencies, and alpha and beta coefficients of
  reordered models.
}
\description{
From subsampling bootstrap results (summary files), produce stability
measures.
}
