% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_mn}
\alias{init_mn}
\title{Initialize the cluster centers (naively).}
\usage{
init_mn(ylist, numclust, TT, dimdat, countslist = NULL)
}
\arguments{
\item{ylist}{A T-length list of (nt  by 3) datasets.  There should  be T of
such datasets. 3 is actually \code{mulen}.}

\item{numclust}{Number of clusters (M).}

\item{TT}{total number of (training) time points.}
}
\value{
An array of dimension (T x dimdat x M).
}
\description{
Initialize the cluster centers (naively).
}
