% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{make_cv_folds}
\alias{make_cv_folds}
\title{Define the time folds cross-validation.}
\usage{
make_cv_folds(ylist = NULL, nfold, blocksize = 20, TT = NULL)
}
\arguments{
\item{blocksize}{Size (number of time indices) of one block.}
}
\value{
List of fold indices.
}
\description{
Define the time folds cross-validation.
}
