% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-data.R
\name{make_grid}
\alias{make_grid}
\title{Helper to make grid. Takes 3-lengthed list of ranges (2 length vectors
containing min and max) and returns a 3-lengthed list of grid points.}
\usage{
make_grid(ylist, gridsize = 5)
}
\arguments{
\item{ylist}{T-length list each containing response matrices of size (nt x
3), which contains coordinates of the 3-variate particles, organized over
time (T) and with (nt) particles at every time.}

\item{gridsize}{Size of equally sized grid, to be formed in each dimension
of data.}
}
\description{
Helper to make grid. Takes 3-lengthed list of ranges (2 length vectors
containing min and max) and returns a 3-lengthed list of grid points.
}
