% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-admm.R
\name{Mstep_beta_admm}
\alias{Mstep_beta_admm}
\title{The M step of beta, using ADMM. (TODO: This should be able to use the
eigendecomp of the Sigmas for the objective value calculation. That is next
up.)}
\usage{
Mstep_beta_admm(
  resp,
  ylist,
  X,
  mean_lambda = 0,
  sigma,
  sigma_eig_by_clust = NULL,
  first_iter = TRUE,
  betas = NULL,
  Zs = NULL,
  Ws = NULL,
  Us = NULL,
  maxdev = NULL,
  niter = 10000,
  rho = 100,
  err_rel = 0.001,
  err_abs = 0,
  zerothresh = 1e-06,
  local_adapt = FALSE,
  local_adapt_niter = 5,
  space = 50
)
}
\arguments{
\item{resp}{Responsibilities; a length-\code{T} list of (\code{nt x K}) responsibility matrices.}

\item{ylist}{T-length list each containing response matrices of size (nt x
d), which contains coordinates of the d-variate particles, organized over
time (T) and with (nt) particles at every time point.}

\item{X}{Matrix of size (T x p+1)}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{sigma}{(numclust x dimdat x dimdat) array.}

\item{sigma_eig_by_clust}{A (numclust x dimdat x dimdat) array; result of running
\code{eigendecomp_sigma_array(sigma)}.}

\item{first_iter}{\code{TRUE} if this is the first EM iteration, which is
handled separately.}

\item{betas}{Warm-start variable.}

\item{Zs}{Warm-start variable.}

\item{Ws}{Warm-start variable.}

\item{Us}{Warm-start variable.}

\item{maxdev}{Radius for maximum deviation of cluster means over time.}

\item{niter}{Total number of outer iterations.}

\item{rho}{todo: fill in}

\item{err_rel}{todo: fill in}

\item{err_abs}{todo: fill in}

\item{zerothresh}{Beta coefficient values below \code{zerothresh} are set to zero.}

\item{local_adapt}{TRUE if locally adaptive ADMM (LA-ADMM) is to be used. If
so, \code{niter} becomes the inner number of iterations, and
\code{local_adapt_niter} becomes the number of outer iterations.}

\item{local_adapt_niter}{Number of outer iterations for LA-ADMM.}

\item{space}{Objectives are only calculated once every \code{space} inner
admm iterations.}
}
\value{
Result of M step; a |numclust| length list of (p+1)x(d) matrices,
  each containing the estimated coefficients for the mean estimation.
}
\description{
The M step of beta, using ADMM. (TODO: This should be able to use the
eigendecomp of the Sigmas for the objective value calculation. That is next
up.)
}
