% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep-proximal.R
\name{Mstep_beta_proximal_with_acceleration}
\alias{Mstep_beta_proximal_with_acceleration}
\title{The M step of beta, using proximal gradient descent. Only to be used when
there is no regularization.}
\usage{
Mstep_beta_proximal_with_acceleration(
  resp,
  ylist,
  X,
  mean_lambda = 0,
  sigma,
  sigma_eig_by_clust = NULL,
  first_iter = TRUE,
  betas = NULL,
  maxdev = NULL,
  niter = 10000,
  err_rel = 0.001,
  err_abs = 0,
  space = 50
)
}
\arguments{
\item{betas}{Warm-start variable.}

\item{niter}{Total number of outer iterations.}

\item{err_rel}{todo: fill in}

\item{err_abs}{todo: fill in}

\item{space}{Objectives are only calculated once every \code{space} inner
admm iterations.}
}
\value{
Result of M step; a |numclust| length list of (p+1)x(d) matrices,
  each containing the estimated coefficients for the mean estimation.
}
\description{
The M step of beta, using proximal gradient descent. Only to be used when
there is no regularization.
}
