% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step_sigma_flowcut.R
\name{Mstep_sigma_flowcut}
\alias{Mstep_sigma_flowcut}
\title{M-step for covariance matrix Sigma_k for clusters k = 1..K}
\usage{
Mstep_sigma_flowcut(
  resp,
  ylist,
  cens_mat,
  first_moment_list,
  second_moment_list,
  mn
)
}
\arguments{
\item{resp}{responsibilities (list length TT of nt x K)}

\item{ylist}{list length K; [[iclust]] is list length TT of nt x d (y_new)}

\item{cens_mat}{censoring matrix with columns (tt, ii, flags...)}

\item{first_moment_list}{list of first conditional moments (by cluster/time/obs)}

\item{second_moment_list}{list of second conditional moments (by cluster/time/obs)}

\item{mn}{fitted means array (TT x d x K)}
}
\value{
(K x d x d) array of covariance estimates
}
\description{
M-step for covariance matrix Sigma_k for clusters k = 1..K
}
