% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mtsqrt_inv}
\alias{mtsqrt_inv}
\title{Given a matrix positive definite matrix a, compute \eqn{a^{-1/2}}.  Only works for
positive semidefinite matrices that are diagonalizable (no normal Jordan
forms, etc.)}
\usage{
mtsqrt_inv(a)
}
\arguments{
\item{a}{A PSD matrix.}
}
\value{
Matrix of the same size as \code{a}.
}
\description{
Given a matrix positive definite matrix a, compute \eqn{a^{-1/2}}.  Only works for
positive semidefinite matrices that are diagonalizable (no normal Jordan
forms, etc.)
}
