% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Objectives (negative log likelihood).}
\usage{
objective(
  mu,
  prob,
  sigma,
  ylist,
  prob_lambda = 0,
  mean_lambda = 0,
  alpha = NULL,
  beta = NULL,
  denslist_by_clust = NULL,
  countslist = NULL,
  each = FALSE,
  sep = FALSE
)
}
\arguments{
\item{mu}{Array of dimension (T x dimdat x numclust ) ; old: T by M by p.}

\item{prob}{Matrix of mixture proportions, T by M.}

\item{sigma}{Array of dimension T by M by p by p.}

\item{ylist}{Data.}

\item{prob_lambda}{lambda for lasso for probabilities.}

\item{mean_lambda}{lambda for lasso for the mean.}

\item{alpha}{Alpha coefficients.}

\item{beta}{Beta coefficients.}

\item{denslist_by_clust}{Pre-calculated densities.}

\item{countslist}{(Optional) Counts corresponding to \code{ylist}.}

\item{each}{if TRUE, return separate log likelihood contributions for each cytogram.}

\item{sep}{if TRUE, return separate log likelihood contributions for each particle.}
}
\value{
Either the original objective value or, T-lengthed vector of
  objective values at each time point.
}
\description{
Objectives (negative log likelihood).
}
