% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix.R
\name{predict.flowmix}
\alias{predict.flowmix}
\title{Prediction: Given new X's, generate a set of means and probs (and return the
same Sigma).}
\usage{
\method{predict}{flowmix}(res, newx = NULL)
}
\arguments{
\item{res}{Object returned from covariate EM flowmix().}

\item{newx}{New covariate.}
}
\value{
List containing mean, prob, and sigma.
}
\description{
Prediction: Given new X's, generate a set of means and probs (and return the
same Sigma).
}
