% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmix.R
\name{predict.flowmix}
\alias{predict.flowmix}
\title{Prediction: Given new covariates X's, generate a set of predicted cluster
means and probs (and return the same Sigma).}
\usage{
\method{predict}{flowmix}(object, ...)
}
\arguments{
\item{object}{Object returned from \code{flowmix()}.}

\item{...}{Make sure to provide new covariate values, a row vector in the
same format and column names as the original X used in \code{object}, in
\code{newx}.}
}
\value{
List containing mean, prob, and sigma.
}
\description{
Prediction: Given new covariates X's, generate a set of predicted cluster
means and probs (and return the same Sigma).
}
